### 2020-05-21 #2199 @whedon
[PRE REVIEW]: NetworkSedimentTransporter: A Landlab submodule for bed material transport through river networks
**Submitting author:** @pfeiffea (<a href="http://orcid.org/0000-0002-3974-132X">Allison Pfeiffer</a>)
**Repository:** <a href="https://github.com/landlab/landlab" target ="_blank">https://github.com/landlab/landlab</a>
**Version:** 2.1
