### 2020-05-21 #2199 @kthyng
@kbarnhart, @openjournals/dev - Do you know the typical setup for reviewing a landlab module? I would like the statistics above to show the number of lines for just this submission (https://github.com/landlab/landlab/tree/master/landlab/components/network_sediment_transporter). Somehow it did find the correct paper — those move to be elsewhere when they aren't being reviewed.
