### 2020-05-21 #2199 @kbarnhart
@kthyng I can provide some info. I first note that I am an author on this submission.

Most Landlab component submissions to JOSS (all, I think) have three parts: component source code, tests, and example notebooks. Each live in a separate location in the Landlab repository. In the case of this specific contribution the content is located at:

- component source: `landlab/components/network_sediment_transporter` and `landlab/plot/network_sediment_transporter`
- tests: `tests/components/network_sediment_transporter`
- example notebooks: `notebooks/tutorials/network_sediment_transporter`

Perhaps run cloc on this set of folders?

re finding the right paper: yes, there is only one paper.md, the prior submissions were renamed to address this issue.

The only thing that is really different about the review process for a landlab component (as compared with a typical submission in which the entire repository is the focus of review)... the reviewers need to be reminded that the scope of the review is not the entire repo. Obviously there  are some whole-package things that get reviewed (install instructions, docs, contributing, etc).

Let me know if there are aspects of your question that I haven't addressed.
