+++ landlab/components/hand_calculator/__init__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python
+"""
+.. codeauthor:: D Litwin
+
+.. sectionauthor:: D Litwin
+"""
+
+from .hand_calculator import HeightAboveDrainage
+
+__all__ = ["HeightAboveDrainage"]
