### 2020-05-26 747bffb
initialize fields, fix indentation
