+++ landlab/components/hand_calculator/hand_calculator.py
@@ -112,47 +112,65 @@ class HeightAboveDrainage(Component):
         self._channel_mask = channel__mask
         self._elev = grid.at_node["topographic__elevation"]
         self._receivers = grid.at_node["flow__receiver_node"]
-        self._downstream_drainage_id = grid.at_node["downstream_drainage__node"]
-        self._hand = grid.at_node["height_above_drainage__elevation"]
+
+        # Downstream drainage node
+        if "downstream_drainage__node" in grid.at_node:
+            self._downstream_drainage_id = grid.at_node[
+                "downstream_drainage__node"
+            ]
+        else:
+            self._downstream_drainage_id = grid.add_zeros(
+                "downstream_drainage__node", at="node", dtype=int
+            )
+
+        # height above nearest drainage
+        if "height_above_drainage__elevation" in grid.at_node:
+            self._hand = grid.at_node[
+                "height_above_drainage__elevation"
+            ]
+        else:
+            self._hand = grid.add_zeros(
+                "height_above_drainage__elevation", at="node", dtype=float
+            )
 
 
-        @property
-        def channel_mask(self):
-            return self._channel_mask
+    @property
+    def channel_mask(self):
+        return self._channel_mask
 
-        @channel_mask.setter
-        def channel_mask(self,new_val):
-            self._channel_mask = new_val
+    @channel_mask.setter
+    def channel_mask(self,new_val):
+        self._channel_mask = new_val
 
 
-        def run_one_step(self):
+    def run_one_step(self):
 
-            self_draining_nodes = sum(self._receivers == np.arange(self._grid.number_of_nodes))
-            if self_draining_nodes != len(self._grid.boundary_nodes):
-                warn(
-                "Pits detected in the flow directions supplied. "
-                "Pits will be treated as drainage nodes."
-                )
+        self_draining_nodes = sum(self._receivers == np.arange(self._grid.number_of_nodes))
+        if self_draining_nodes != len(self._grid.boundary_nodes):
+            warn(
+            "Pits detected in the flow directions supplied. "
+            "Pits will be treated as drainage nodes."
+            )
 
-            self._downstream_drainage_id[:] = 0
-            is_drainage_node = self._channel_mask
-            is_drainage_node[self._grid.open_boundary_nodes] = 1
+        self._downstream_drainage_id[:] = 0
+        is_drainage_node = self._channel_mask
+        is_drainage_node[self._grid.open_boundary_nodes] = 1
 
-            for i in range(self._grid.number_of_nodes):
+        for i in range(self._grid.number_of_nodes):
 
-                if i == receivers[i] or is_drainage_node[i]: #started on a boundary, depression, or in channel
-                    self._downstream_drainage_id[i] = i
+            if i == self._receivers[i] or is_drainage_node[i]: #started on a boundary, depression, or in channel
+                self._downstream_drainage_id[i] = i
 
-                else:
-                    cur_node = i
-                    reached_drainage = False
-                    while not reached_drainage:
-                        downstream_id = self._receivers[cur_node]
-                        if is_drainage_node[downstream_id]:
-                            self._downstream_drainage_id[i] = downstream_id
-                            reached_drainage = True
-                        else:
-                            cur_node = downstream_id
+            else:
+                cur_node = i
+                reached_drainage = False
+                while not reached_drainage:
+                    downstream_id = self._receivers[cur_node]
+                    if is_drainage_node[downstream_id]:
+                        self._downstream_drainage_id[i] = downstream_id
+                        reached_drainage = True
+                    else:
+                        cur_node = downstream_id
 
-            nearest_drainage_elev = self._elev[self._downstream_drainage_id]
[…]
