### 2020-05-26 43a3063
Start hand tests
