### 2020-05-26 d023957
Add description, change pit handling
