### 2020-05-27 cd8f6c6
change name and docstring
