+++ tests/components/hand_calculator/test_hand.py
@@ -3,7 +3,7 @@ import pytest
 from numpy.testing import assert_equal
 
 from landlab import FieldError, RasterModelGrid
-from landlab.components import HeightAboveDrainage, FlowAccumulator
+from landlab.components import HeightAboveDrainageCalculator, FlowAccumulator
 
 
 def test_route_to_multiple_error_raised():
@@ -16,7 +16,7 @@ def test_route_to_multiple_error_raised():
     channel__mask = mg.zeros(at="node")
 
     with pytest.raises(NotImplementedError):
-        HeightAboveDrainage(mg, channel__mask)
+        HeightAboveDrainageCalculator(mg, channel__mask)
 
 def test_warn_drainage_pits():
     mg = RasterModelGrid((4, 4))
@@ -29,7 +29,7 @@ def test_warn_drainage_pits():
 
     channel__mask = mg.zeros(at="node")
     channel__mask[[2,6]] = 1
-    hd = HeightAboveDrainage(mg, channel__mask)
+    hd = HeightAboveDrainageCalculator(mg, channel__mask)
 
     with pytest.warns(UserWarning):
         hd.run_one_step()
