### 2020-05-31 a743154
fix components init and add parameter description
