+++ landlab/components/hand_calculator/hand_calculator.py
@@ -83,6 +83,14 @@ class HeightAboveDrainageCalculator(Component):
     _unit_agnostic = True
 
     _info = {
+        "channel__mask": {
+            "dtype": np.uint8,
+            "intent": "in",
+            "optional": True,
+            "units": "-",
+            "mapping": "node",
+            "doc": "Logical map of at which grid nodes channels are present",
+        },
         "flow__receiver_node": {
             "dtype": int,
             "intent": "in",
