### 2020-06-02 #2199 @kthyng
@openjournals/dev Might it be possible in the future to run `@whedon check repository` on a subset of a repository, sort of like generating a pdf from a branch of a repo? Just an idea.
