### 2020-06-02 #2199 @kthyng
For reference, I am going to paste in the `cloc` output I ran on my laptop since the information above is not accurate for this specific project:

```
(base) tatoosh:~ kthyng$ cloc Downloads/landlab-master/landlab/components/network_sediment_transporter
       3 text files.
       3 unique files.
       0 files ignored.

Language                     files          blank        comment           code
Python                           2            229            426            545
Markdown                         1             10              0             17
SUM:                             3            239            426            562
(base) tatoosh:~ kthyng$ cloc Downloads/landlab-master/landlab/plot/network_sediment_transporter
       3 text files.
       3 unique files.
       0 files ignored.

Language                     files          blank        comment           code
Python                           3             99            187            301
SUM:                             3             99            187            301
(base) tatoosh:~ kthyng$ cloc Downloads/landlab-master/tests/components/network_sediment_transporter
       8 text files.
       8 unique files.
       1 file ignored.

Language                     files          blank        comment           code
Python                           8            173             20            599
SUM:                             8            173             20            599
(base) tatoosh:~ kthyng$ cloc Downloads/landlab-master/notebooks/tutorials/network_sediment_transporter
       4 text files.
       4 unique files.
       0 files ignored.

Language                     files          blank        comment           code
Jupyter Notebook                 4              0           1298            664
SUM:                             4              0           1298            664
```
