### 2020-06-02 #2199 @kthyng
@pfeiffea  I just read your personal description on github and it made me happy.

Sorry for my delay, getting work done right now is slow going. Could you look at this list and give a handful of suggestions for reviewers? Please don't @ them so that they aren't specifically mentioned, just give their github handle, like kthyng for me. https://bit.ly/joss-reviewers
