### 2020-06-03 f4d4fc0
fix channel mask in init
