### 2020-06-03 #2199 @arfon
> @openjournals/dev Might it be possible in the future to run @whedon check repository on a subset of a repository, sort of like generating a pdf from a branch of a repo? Just an idea.

Good idea. I've added this as a possible enhancement in https://github.com/openjournals/whedon-api/issues/101
