### 2020-06-03 27d4a09
Delint tests
