### 2020-06-04 d7cc1a9
Add HeightAboveDrainageCalculator component (#1207)
* Begin HeightAboveDrainage component

* Add warn

* initialize fields, fix indentation

* delint

* Start hand tests

* Add description, change pit handling

* change name and docstring

* Finish update name

* Fix docstring tests?

* Use stack

* delint

* Fix optional field

* fix components init and add parameter description

* Channel mask info

* fix channel mask in init

* delint

* Delint tests

* little stuff
