### 2020-06-05 63d623a
eliminate duplicate calling of stack-construction fns
