### 2020-06-05 25764b7
enable FlowAccumulator to accept a LakeMapperBarnes as alternative to DepressionFinderAndRouter
