### 2020-06-05 710f909
create output fields before other compos
