+++ landlab/components/flow_accum/flow_accumulator.py
@@ -690,7 +690,6 @@ class FlowAccumulator(Component):
         keyword arguments, tests the argument of runoff_rate, and
         initializes new fields.
         """
-        # print('FA>init')
         super().__init__(grid)
         # Keep a local reference to the grid
 
@@ -723,12 +722,7 @@ class FlowAccumulator(Component):
 
         self._node_cell_area = node_cell_area
 
-        # STEP 2:
-        # identify Flow Director method, save name, import and initialize the correct
-        # flow director component if necessary
-        self._add_director(flow_director)
-        self._add_depression_finder(depression_finder)
-
+        # STEP 3:
         # This component will track of the following variables.
         # Attempt to create each, if they already exist, assign the existing
         # version to the local copy.
@@ -752,6 +746,13 @@ class FlowAccumulator(Component):
 
         self._D_structure = self._grid.BAD_INDEX * grid.ones(at="link", dtype=int)
         self._nodes_not_in_stack = True
+
+        # STEP 4:
+        # identify Flow Director method, save name, import and initialize the 
+        # correct flow director component if necessary; same with 
+        # lake/depression handler, if specified.
+        self._add_director(flow_director)
+        self._add_depression_finder(depression_finder)
 
         if len(self._kwargs) > 0:
             kwdstr = " ".join(list(self._kwargs.keys()))
