### 2020-06-05 366a826
add test with LakeMapperBarnes
