+++ tests/components/hand_calculator/test_hand.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from landlab import RasterModelGrid
-from landlab.components import HeightAboveDrainageCalculator, FlowAccumulator
+from landlab.components import FlowAccumulator, HeightAboveDrainageCalculator
 
 
 def test_route_to_multiple_error_raised():
