### 2020-06-05 b31b15f
add tests for specifying LMB as class or object
