+++ tests/components/flow_accum/test_flow_accumulator.py
@@ -1210,3 +1210,8 @@ def test_with_lake_mapper_barnes():
 
     rcvrs = hmg_hole.at_node["flow__receiver_node"]
     assert_array_equal(rcvrs[[13, 21, 46]], [6, 13, 39])
+
+    from landlab.components import LakeMapperBarnes
+    fa = FlowAccumulator(hmg_hole, depression_finder=LakeMapperBarnes)
+    lmb = LakeMapperBarnes(hmg_hole)
+    fa = FlowAccumulator(hmg_hole, depression_finder=lmb)
