+++ landlab/grid/landslide_unitgrid.py
@@ -6,25 +6,42 @@ import pytest
 from numpy.testing import assert_array_almost_equal
 
 from landlab import RasterModelGrid
-from landlab.components import LandslideProbability
 
+(_SHAPE, _SPACING, _ORIGIN) = ((5, 4), (10e0, 10e0), (0.0, 0.0))
+_ARGS = (_SHAPE, _SPACING, _ORIGIN)
+
+rowcol = _SHAPE 
+edgelength = _SPACING
 relative_wetness = 0.75
 
-def unitgrid(dimensions):
-    shape = (5, 4)
-    spacing = (10e0, 10e0)
-    coordinates = (0.0, 0.0)
-    grid = RasterModelGrid(shape,spacing,coordinates)
-    gridnum=grid.number_of_nodes
-    
-    assert gridnum == dimensions[0][0]*dimensions[0][1]
-    
-    return grid, gridnum
+unit_default_value={}
+unit_default_value["topographic__slope"]=(0.1,0.8)
+unit_default_value["topographic__specific_contributing_area"]=(10,100)
+unit_default_value["soil__transmissivity"]=(0.1, 10)
+unit_default_value["soil__mode_total_cohesion"]=(30, 900)
+mode = unit_default_value["soil__mode_total_cohesion"]
+# code used to run the model uses a random scatter <cohesion_scatter = np.random.randint(1,10)>
+scatter = np.random.randint(1,10)  #smallest min is the largest scatter, minus the mode cohesion
+unit_default_value["soil__minimum_total_cohesion"]=((mode[0]-scatter),mode[1]-scatter)
+#large max is the largest scatter, plus the mode cohesion
+unit_default_value["soil__maximum_total_cohesion"]=((mode[0]+scatter),mode[1]+scatter)
+unit_default_value["soil__internal_friction_angle"]=(26, 37)
+unit_default_value["soil__thickness"]=(0.1,1) #functionally setting to uniform uniform 1 meter default; coded for the option of a range of values
+unit_default_value["soil__density"]=(1999,2001) #functionally setting to uniform 2000
 
-def calc_unit_default_value():
+#Note on Ksat default range calculation
+#reverse order: (see above) lowest Transmissivity value are in upper right of synthethic grid,
+#not reversed: (see above) highest soil thickness in upper right of synthethic grid,
+#caution creating Ksat since parameterization of T and hs will have inverse relationship on Ksat
+#Kmin  = Tmin/hs_max   -slow- think of a deep thick clay layer 
+kmin = unit_default_value['soil__transmissivity'][0]/unit_default_value['soil__thickness'][1] 
+#Kmax = Tmax/hs_min    -fast- think of a sieve or a parking lot
+kmax = unit_default_value['soil__transmissivity'][1]/unit_default_value['soil__thickness'][0] 
+unit_default_value["soil__saturated_hydraulic_conductivity"]= (kmin,kmax)
+
+def unit_defaultpar_value():
 
     unit_default_value={}
-    scatter = np.random.randint(1,10)  #smallest min is the largest scatter, minus the mode cohesion
     unit_default_value["topographic__slope"]=(0.1,0.8)
     unit_default_value["topographic__specific_contributing_area"]=(10,100)
     unit_default_value["soil__transmissivity"]=(0.1, 10)
@@ -54,197 +71,16 @@ def calc_unit_default_value():
     unit_default_value["soil__mean_watertable_depth"]=(0,1000)
     unit_default_value["soil__mean_recharge"]=(0,1000)
     unit_default_value["soil__probability_of_saturation"]=(0,1)
+
     return unit_default_value
 
-def get_default_nodevalues():
-    """Print out the input range for
-    all landslide component values (min/max).
-    ----------
[…]
