+++ tests/components/landslides/conftest.py
@@ -1,21 +1,189 @@
 import pytest
-
+import numpy as np
 from landlab import RasterModelGrid
 from landlab.components import LandslideProbability
 
+@pytest.fixture
+def spacing():
+    spacing = (10e0, 10e0)
+    return spacing
+
+@pytest.fixture
+def shape():
+    shape = (5, 4)
+    return shape
+@pytest.fixture
+def dimensions():
+    dimensions = ((5, 4), (10e0, 10e0), (0.0, 0.0))
+    return dimensions
+    
+@pytest.fixture
+def ls_prob_zeros():
+    dimensions = ((5, 4), (10e0, 10e0), (0.0, 0.0))
+    shape = dimensions[0]
+    spacing = dimensions[1]
+    coordinates = dimensions[2]
+    
+    grid = RasterModelGrid(shape=shape,xy_spacing=spacing,xy_of_lower_left=coordinates)
+    grid.add_zeros("topographic__slope", at="node", dtype=float)
+    grid.add_zeros("topographic__specific_contributing_area", at="node", dtype=int)
+    grid.add_zeros("soil__transmissivity", at="node", dtype=float)
+    grid.add_zeros("soil__saturated_hydraulic_conductivity", at="node", dtype=float)
+    grid.add_zeros("soil__mode_total_cohesion", at="node", dtype=int)
+    grid.add_zeros("soil__minimum_total_cohesion", at="node", dtype=int)
+    grid.add_zeros("soil__maximum_total_cohesion", at="node", dtype=int)
+    grid.add_zeros("soil__internal_friction_angle", at="node", dtype=int)
+    grid.add_zeros("soil__density", at="node", dtype=int)
+    grid.add_zeros("soil__thickness", at="node", dtype=float)
+
+    return LandslideProbability(grid)
+
+@pytest.fixture
+def unit_default_value():
+
+    unit_default_value={}
+    unit_default_value["topographic__slope"]=(0.1,0.8)
+    unit_default_value["topographic__specific_contributing_area"]=(10,100)
+    unit_default_value["soil__transmissivity"]=(0.1, 10)
+    unit_default_value["soil__mode_total_cohesion"]=(30, 900)
+    mode = unit_default_value["soil__mode_total_cohesion"]
+    # code used to run the model uses a random scatter <cohesion_scatter = np.random.randint(1,10)>
+    scatter = np.random.randint(1,10)  #smallest min is the largest scatter, minus the mode cohesion
+    unit_default_value["soil__minimum_total_cohesion"]=((mode[0]-scatter),mode[1]-scatter)
+    #large max is the largest scatter, plus the mode cohesion
+    unit_default_value["soil__maximum_total_cohesion"]=((mode[0]+scatter),mode[1]+scatter)
+    unit_default_value["soil__internal_friction_angle"]=(26, 37)
+    unit_default_value["soil__thickness"]=(0.1,1) #functionally setting to uniform uniform 1 meter default; coded for the option of a range of values
+    unit_default_value["soil__density"]=(1999,2001) #functionally setting to uniform 2000
+    #Note on Ksat default range calculation
+    #reverse order: (see above) lowest Transmissivity value are in upper right of synthethic grid,
+    #not reversed: (see above) highest soil thickness in upper right of synthethic grid,
+    #caution creating Ksat since parameterization of T and hs will have inverse relationship on Ksat
+    #Kmin  = Tmin/hs_max   -slow- think of a deep thick clay layer 
+    kmin = unit_default_value['soil__transmissivity'][0]/unit_default_value['soil__thickness'][1] 
+    #Kmax = Tmax/hs_min    -fast- think of a sieve or a parking lot
[…]
