### 2020-06-08 f99d93c
check for pits before calling lake/depression mapper; make pretty
