### 2020-06-09 55d6244
Start callback function in gdp
