### 2020-06-10 8170de0
update docs, make test
