### 2020-06-10 34697d5
misc docs and handle divide by zero
