### 2020-06-12 #2199 @kthyng
Hi @zsylvester and @ebgoldstein! Are you willing and interested in reviewing this submission to the Journal for Open Source Software? The reviews are all done openly through GitHub and given the present world circumstances, I'd hope for reviews within about 6 weeks. You can learn more about the review process here: https://joss.readthedocs.io/en/latest/review_criteria.html
