### 2020-06-12 #2199 @zsylvester
Hi @kthyng - I will need to learn a bit about the JOSS review process, but yes, I am interested.
