### 2020-06-13 e1d444a
Make docs and links for hand_calculator
