+++ docs/source/reference/components/hand_calculator.rst
@@ -1,0 +1,7 @@
+HeightAboveDrainageCalculator: Calculate height above nearest drainage
+-------------------------------------------------
+
+.. automodule:: landlab.components.hand_calculator.hand_calculator
+    :members:
+    :undoc-members:
+    :show-inheritance:
