### 2020-06-15 #2199 @kthyng
Hi @zsylvester and @ebgoldstein: awesome! I will start the review issue and close this one — the rest of the action should take place in the review issue. Please reach out if you have questions on the process. The JOSS docs have information, and it can also help to look through old submissions to see how things work. Thanks!!
