+++ setup.cfg
@@ -40,4 +40,5 @@ ignore =
 	E203
 	E501
 	W503
+  W605 
 max-line-length = 88
