### 2020-06-17 62520fc
remove error exemption
