+++ setup.cfg
@@ -40,5 +40,4 @@ ignore =
 	E203
 	E501
 	W503
-  W605 
 max-line-length = 88
