+++ docs/source/reference/components/hand_calculator.rst
@@ -1,5 +1,5 @@
 HeightAboveDrainageCalculator: Calculate height above nearest drainage
--------------------------------------------------
+----------------------------------------------------------------------
 
 .. automodule:: landlab.components.hand_calculator.hand_calculator
     :members:
