+++ docs/source/user_guide/papers_presentations.rst
@@ -7,6 +7,23 @@ Papers and Presentations
 Peer-reviewed journal articles
 ------------------------------
 
+2020
+````
+
+Barnhart, K.R., Hutton, E.W.H., Tucker, G.E., Gasparini, N.M., Istanbulluoglu, E., Hobley, D.E.J., Lyons⁠, N.J., Mouchene, M., Nudurupati, S.S., Adams, J.M., and Bandaragoda, C. (2020) **Short communication: Landlab 2.0: A software package for Earth surface dynamics.** *Earth Surface Dynamics,* 8, 379–397, `doi:10.5194/esurf-8-379-2020 <https://doi.org/10.5194/esurf-8-379-2020>`__.
+
+Barnhart, K. R., Tucker, G. E., Doty, S., Shobe, C. M., Glade, R. C., Rossi, M. W., & Hill, M. C. (2020). **Inverting topography for landscape evolution model process representation: Part 1, conceptualization and sensitivity analysis.** *Journal of Geophysical Research: Earth Surface,* e2018JF004961. `https://doi.org/10.1029/2018JF004961 <https://doi.org/10.1029/2018JF004961>`__.
+
+Barnhart, K. R., Tucker, G. E., Doty, S., Shobe, C. M., Glade, R. C., Rossi, M. W., & Hill, M. C. (2020). **Inverting topography for landscape evolution model process representation: Part 2, calibration and validation.** *Journal of Geophysical Research: Earth Surface,* e2018JF004963. `https://doi.org/10.1029/2018JF004963 < https://doi.org/10.1029/2018JF004963>`__.
+
+Barnhart, K. R., Tucker, G. E., Doty, S., Shobe, C. M., Glade, R. C., Rossi, M. W., & Hill, M. C. (2020). **Inverting topography for landscape evolution model process representation: Part 3, Determining parameter ranges for select mature geomorphic transport laws and connecting changes in fluvial erodibility to changes in climate.** *Journal of Geophysical Research: Earth Surface,* e2019JF005287, `https://doi.org/10.1029/2019JF005287 <https://doi.org/10.1029/2019JF005287>`__.
+
+Carriere, A., Le Bouteiller, C., Tucker, G.E., Klotz, S., and Naaim, M. (2020) **Impact of vegetation on erosion: Insights from the calibration and test of a landscape evolution model in alpine badland catchments.** *Earth Surface Processes and Landforms.* `https://doi.org/10.1002/esp.4741 <https://doi.org/10.1002/esp.4741>`__.
+
+Litwin, D.G., Tucker, G.E., Barnhart, K.R., and Harman, C.J. (2020) **GroundwaterDupuitPercolator: A Landlab component for groundwater flow,** *Journal of Open Source Software,* 5(46), 1935, `https://doi.org/10.21105/joss.01935 <https://doi.org/10.21105/joss.01935>`__.
+
+Tucker, G. E., Hobley, D.E.J., McCoy, S.W., and Struble, W.T. (2020) **Modeling the shape and evolution of normal-fault facets.** *Journal of Geophysical Research: Earth Surface,* 125, `https://doi.org/10.1029/2019JF005305 <https://doi.org/10.1029/2019JF005305>`__.
+
 2019
 ````
 
@@ -31,7 +48,7 @@ feedbacks.** *Geology*,
 `doi:10.1130/G46219.1 <https://pubs.geoscienceworld.org/gsa/geology/article/47/7/650/570313/Canyon-shape-and-erosion-dynamics-governed-by>`__. \*Equal
 contributions
 
[…]
