+++ requirements.txt
@@ -1,9 +1,0 @@
-bmipy<1
-matplotlib
-netcdf4
-pyyaml
-pyshp
-scipy
-statsmodels
-pandas
-xarray
