### 2020-06-22 d7af1d0
removed quickstart.md
