+++ docs/source/getting_started/quickstart.md
@@ -1,1 +1,0 @@
-notes and practice
