### 2020-06-27 469a670
remove landslide link in readme.rst
