+++ README.rst
@@ -19,9 +19,6 @@
 .. image:: https://mybinder.org/badge_logo.svg
  :target: https://mybinder.org/v2/gh/landlab/landlab/release?filepath=notebooks/welcome.ipynb
  
- ## Click here to Launch Binder to use the Version 2 Unit Test Interactive Notebook
-.. image:: https://mybinder.org/badge_logo.svg
- :target: https://mybinder.org/v2/gh/ChristinaB/landlab/ChristinaB-landslide?filepath=notebooks%2Ftutorials%2Flandslides%2FUnit_LandlabLandslide_v2-2020430.ipynb
 =======
 Landlab
 =======
