### 2020-06-27 1d69d46
fix README.rst with git reset
