+++ landlab/components/landslides/landslide_probability.py
@@ -31,13 +31,15 @@ Add depth to water table as an optional hydrology input (alternative to recharge
 """
 
 import copy
-import pandas as pd
+
 import numpy as np
+import pandas as pd
 import scipy.constants
 from scipy import interpolate
 from statsmodels.distributions.empirical_distribution import ECDF
 
 from landlab import Component
+
 
 class LandslideProbability(Component):
     """Landslide probability component using the infinite slope stability
@@ -48,15 +50,15 @@ class LandslideProbability(Component):
     stability index (Factor of Safety).
 
     The driving force for failure is provided by the user in the form of
-    groundwater recharge OR depth to groundwater. Four options for providing 
+    groundwater recharge OR depth to groundwater. Four options for providing
     recharge plus four options of depth to groundwater are supported.
     The model uses topographic and soil characteristics provided as input
     by the user.
 
     The main method of the LandslideProbability class is
     `calculate_landslide_probability()`, which calculates the mean soil
-    relative wetness, mean water table depth, mean recharge, probability of soil saturation, 
-    and probability of failure at each node based on a Monte Carlo simulation.   
+    relative wetness, mean water table depth, mean recharge, probability of soil saturation,
+    and probability of failure at each node based on a Monte Carlo simulation.
 
     **Usage:**
 
@@ -103,48 +105,49 @@ class LandslideProbability(Component):
     Option 5 - Uniform depth to groundwater
 
     .. code-block:: python
-    
-        LandslideProbability(grid, 
+
+        LandslideProbability(grid,
                              number_of_iterations=250,
-                             groundwater__depth_distribution='uniform', 
-                             groundwater__depth_min_value=0.01, 
+                             groundwater__depth_distribution='uniform',
+                             groundwater__depth_min_value=0.01,
                              groundwater__depth_max_value=1.5)
-                             
-    Option 6 - Lognormal depth to groundwater 
-    
+
+    Option 6 - Lognormal depth to groundwater
+
     .. code-block:: python
-        
+
         LandslideProbability(grid, number_of_iterations=250,
-                             groundwater__depth_distribution='lognormal', 
+                             groundwater__depth_distribution='lognormal',
                              groundwater__depth_mean=0.5.,
                              groundwater__depth_standard_deviation=0.1)
-                             
-    Option 7 - Lognormal_spatial depth to groundwater 
-    
+
+    Option 7 - Lognormal_spatial depth to groundwater
+
     .. code-block:: python
-    
+
         LandslideProbability(grid, number_of_iterations=250,
-                             groundwater__depth_distribution='lognormal_spatial', 
+                             groundwater__depth_distribution='lognormal_spatial',
                              groundwater__depth_mean=np.random.randint(0, 2, grid_size),
-                             groundwater__depth_standard_deviation=np.random.rand(grid_size))   
-                             
+                             groundwater__depth_standard_deviation=np.random.rand(grid_size))
+
     Option 8 - Data_driven_spatial depth to groundwater
-    
+
     .. code-block:: python
-    
-        LandslideProbability(grid, number_of_iterations=250, 
-                             groundwater__depth_distribution='data_driven_spatial', 
+
+        LandslideProbability(grid, number_of_iterations=250,
+                             groundwater__depth_distribution='data_driven_spatial',
                              groundwater__depth_HSD_inputs={HSD_dict})
 
     Examples
     --------
     >>> from landlab import RasterModelGrid
[…]

[…]
