+++ landlab/components/space/space.py
@@ -7,7 +7,7 @@ from landlab.components.erosion_deposition.generalized_erosion_deposition import
 )
 from landlab.utils.return_array import return_array_at_node
 
-from .cfuncs import calculate_qs_in
+from .cfuncs import calculate_qs_in, calc_erosion_rates
 
 ROOT2 = np.sqrt(2.0)  # syntactic sugar for precalculated square root of 2
 TIME_STEP_FACTOR = 0.5  # factor used in simple subdivision solver
@@ -327,6 +327,8 @@ class Space(_GeneralizedErosionDeposition):
                 self._topographic__elevation - self._soil__depth
             )
 
+        self._sed_erosion_term  = np.zeros(grid.number_of_nodes)
+        self._br_erosion_term  = np.zeros(grid.number_of_nodes)
         self._Es = np.zeros(grid.number_of_nodes)
         self._Er = np.zeros(grid.number_of_nodes)
 
@@ -344,6 +346,11 @@ class Space(_GeneralizedErosionDeposition):
         elif solver == "adaptive":
             self.run_one_step = self.run_with_adaptive_time_step_solver
             self._time_to_flat = np.zeros(grid.number_of_nodes)
+        elif solver == "experimental":
+            self.run_one_step = self.run_with_experimental_adaptive_time_step_solver
+            self._time_to_flat = np.zeros(grid.number_of_nodes)
+            self._time_to_zero_alluv = np.zeros(grid.number_of_nodes)
+            self._dzdt = np.zeros(grid.number_of_nodes)
         else:
             raise ValueError(
                 "Parameter 'solver' must be one of: " + "'basic', 'adaptive'"
@@ -370,8 +377,12 @@ class Space(_GeneralizedErosionDeposition):
     def _calc_erosion_rates(self):
         """Calculate erosion rates."""
         # if sp_crits are zero, then this colapses to correct all the time.
-        omega_sed = self._K_sed * self._Q_to_the_m * np.power(self._slope, self._n_sp)
-        omega_br = self._K_br * self._Q_to_the_m * np.power(self._slope, self._n_sp)
+        if self._n_sp == 1.0:
+            S_to_the_n = self._slope
+        else:
+            S_to_the_n = np.power(self._slope, self._n_sp)
+        omega_sed = self._K_sed * self._Q_to_the_m * S_to_the_n
+        omega_br = self._K_br * self._Q_to_the_m * S_to_the_n
 
         omega_sed_over_sp_crit = np.divide(
             omega_sed,
@@ -397,11 +408,12 @@ class Space(_GeneralizedErosionDeposition):
             1.0 - np.exp(-omega_br_over_sp_crit)
         )
 
+        H_over_Hstar = self._soil__depth / self._H_star
         self._Es = self._sed_erosion_term * (
-            1.0 - np.exp(-self._soil__depth / self._H_star)
+            1.0 - np.exp(-H_over_Hstar)
         )
 
-        self._Er = self._br_erosion_term * np.exp(-self._soil__depth / self._H_star)
+        self._Er = self._br_erosion_term * np.exp(-H_over_Hstar)
 
     @property
     def Es(self):
@@ -419,10 +431,31 @@ class Space(_GeneralizedErosionDeposition):
         return self._H
 
     def _calc_qs_in_and_depo_rate(self):
+        import time
+
         # Choose a method for calculating erosion:
         self._calc_hydrology()
+#        st = time.time()
         self._calc_erosion_rates()
+        # calc_erosion_rates(
+        #     self._K_sed,
+        #     self._K_br,
+        #     self._Q_to_the_m,
+        #     self._slope,
+        #     self._soil__depth,
+        #     self._sed_erosion_term,
[…]
