### 2020-06-28 b612217
remove test cy implementation
