+++ landlab/components/space/space.py
@@ -7,7 +7,7 @@ from landlab.components.erosion_deposition.generalized_erosion_deposition import
 )
 from landlab.utils.return_array import return_array_at_node
 
-from .cfuncs import calculate_qs_in, calc_erosion_rates
+from .cfuncs import calculate_qs_in
 
 ROOT2 = np.sqrt(2.0)  # syntactic sugar for precalculated square root of 2
 TIME_STEP_FACTOR = 0.5  # factor used in simple subdivision solver
@@ -346,9 +346,6 @@ class Space(_GeneralizedErosionDeposition):
         elif solver == "adaptive":
             self.run_one_step = self.run_with_adaptive_time_step_solver
             self._time_to_flat = np.zeros(grid.number_of_nodes)
-        elif solver == "experimental":
-            self.run_one_step = self.run_with_experimental_adaptive_time_step_solver
-            self._time_to_flat = np.zeros(grid.number_of_nodes)
             self._time_to_zero_alluv = np.zeros(grid.number_of_nodes)
             self._dzdt = np.zeros(grid.number_of_nodes)
         else:
@@ -431,31 +428,11 @@ class Space(_GeneralizedErosionDeposition):
         return self._H
 
     def _calc_qs_in_and_depo_rate(self):
-        import time
 
         # Choose a method for calculating erosion:
         self._calc_hydrology()
-#        st = time.time()
         self._calc_erosion_rates()
-        # calc_erosion_rates(
-        #     self._K_sed,
-        #     self._K_br,
-        #     self._Q_to_the_m,
-        #     self._slope,
-        #     self._soil__depth,
-        #     self._sed_erosion_term,
-        #     self._br_erosion_term,
-        #     self._Es,
-        #     self._Er,
-        #     self._sp_crit_sed,
-        #     self._sp_crit_br,
-        #     self._n_sp,
-        #     self._H_star,
-        #     self._phi,
-        # )
 
-#        en = time.time()
-#        print(en-st)
         is_flooded_core_node = self._get_flooded_core_nodes()
 
         self._Es[is_flooded_core_node] = 0.0
@@ -623,124 +600,6 @@ class Space(_GeneralizedErosionDeposition):
         )
 
     def run_with_adaptive_time_step_solver(self, dt=1.0):
-        """Run step with CHILD-like solver that adjusts time steps to prevent
-        slope flattening.
-
-        Parameters
-        ----------
-        dt : float
-            Model timestep [T]
-
-        Examples
-        --------
-        >>> from landlab import RasterModelGrid
-        >>> from landlab.components import FlowAccumulator
-        >>> import numpy as np
-
-        >>> rg = RasterModelGrid((3, 4))
-        >>> z = rg.add_zeros('topographic__elevation', at='node')
-        >>> z[:] = 0.1 * rg.x_of_node
-        >>> H = rg.add_zeros('soil__depth', at='node')
-        >>> H += 0.1
-        >>> br = rg.add_zeros('bedrock__elevation', at='node')
-        >>> br[:] = z - H
-
-        >>> fa = FlowAccumulator(rg, flow_director='FlowDirectorSteepest')
-        >>> fa.run_one_step()
-        >>> sp = Space(rg, K_sed=1.0, K_br=0.1,
-        ...            F_f=0.5, phi=0.0, H_star=1., v_s=1.0,
-        ...            m_sp=0.5, n_sp = 1.0, sp_crit_sed=0,
-        ...            sp_crit_br=0, solver='adaptive')
-        >>> sp.run_one_step(dt=10.0)
-
-        >>> np.round(sp.Es[5:7], 4)
-        array([ 0.0029,  0.0074])
-        >>> np.round(sp.Er[5:7], 4)
[…]
