### 2020-07-04 20be337
add properties for core-core, fv-core, and core-fv links
