### 2020-07-04 c352af8
add matrix creation utility 1st draft
