### 2020-07-07 1d2ee0f
add fns to __init__
