### 2020-07-10 d0f44ea
add general link_by_tail_and_head_status method
