### 2020-07-10 578d042
convert to list for doctest to avoid silly test fail around reporting or not the dtype of array
