+++ landlab/grid/base.py
@@ -987,62 +987,6 @@ class ModelGrid(GraphFields, EventLayersMixIn, MaterialLayersMixIn):
         """
         return np.where(self.status_at_link == LinkStatus.FIXED)[0]
 
-    @property
-    @cache_result_in_object()
-    @return_readonly_id_array
-    def core_to_core_links(self):
-        """Return an array with the IDs of all links that join two core nodes.
-
-        Examples
-        --------
-        >>> from landlab import RasterModelGrid
-        >>> grid = RasterModelGrid((4, 5))
-        >>> grid.status_at_node[13] = grid.BC_NODE_IS_FIXED_VALUE
-        >>> grid.status_at_node[2] = grid.BC_NODE_IS_CLOSED
-        >>> grid.core_to_core_links
-        array([10, 11, 14, 15, 19])
-        """
-        return self.links_by_tail_and_head_status(self.BC_NODE_IS_CORE,
-                                                  self.BC_NODE_IS_CORE)
-
-    @property
-    @cache_result_in_object()
-    @return_readonly_id_array
-    def core_to_fixed_value_links(self):
-        """Return an array with the IDs of all links that a core node (tail)
-        to a fixed-value node (head).
-
-        Examples
-        --------
-        >>> from landlab import RasterModelGrid
-        >>> grid = RasterModelGrid((4, 5))
-        >>> grid.status_at_node[13] = grid.BC_NODE_IS_FIXED_VALUE
-        >>> grid.status_at_node[2] = grid.BC_NODE_IS_CLOSED
-        >>> grid.core_to_fixed_value_links
-        array([12, 16, 20, 23, 24])
-        """
-        return self.links_by_tail_and_head_status(self.BC_NODE_IS_CORE,
-                                                  self.BC_NODE_IS_FIXED_VALUE)
-
-    @property
-    @cache_result_in_object()
-    @return_readonly_id_array
-    def fixed_value_to_core_links(self):
-        """Return an array with the IDs of all links that join a fixed-value node (tail)
-        to a core node (head).
-
-        Examples
-        --------
-        >>> from landlab import RasterModelGrid
-        >>> grid = RasterModelGrid((4, 5))
-        >>> grid.status_at_node[13] = grid.BC_NODE_IS_FIXED_VALUE
-        >>> grid.status_at_node[2] = grid.BC_NODE_IS_CLOSED
-        >>> grid.fixed_value_to_core_links
-        array([ 5,  7,  9, 18])
-        """
-        return self.links_by_tail_and_head_status(self.BC_NODE_IS_FIXED_VALUE,
-                                                  self.BC_NODE_IS_CORE)
-
     def links_by_tail_and_head_status(self, status_at_tail, status_at_head):
         """Return an array with the IDs of all links that join a fixed-value node (tail)
         to a core node (head).
@@ -1055,12 +999,12 @@ class ModelGrid(GraphFields, EventLayersMixIn, MaterialLayersMixIn):
         >>> core = grid.BC_NODE_IS_CORE
         >>> grid.status_at_node[13] = fv
         >>> grid.status_at_node[2] = grid.BC_NODE_IS_CLOSED
-        >>> grid.links_by_tail_and_head_status(core, core)
-        array([10, 11, 14, 15, 19])
-        >>> grid.links_by_tail_and_head_status(core, fv)
-        array([12, 16, 20, 23, 24])
-        >>> grid.links_by_tail_and_head_status(fv, core)
-        array([ 5,  7,  9, 18])
+        >>> grid.links_by_tail_and_head_status(core, core).tolist()
+        [10, 11, 14, 15, 19]
+        >>> grid.links_by_tail_and_head_status(core, fv).tolist()
[…]
