### 2020-07-10 f6671df
update to use generic links_by_tail_and_head_status
