### 2020-07-10 6b2c08e
correct docstring
