### 2020-07-13 #2341 @ebgoldstein
Hi @pfeiffea — I'm walking through the tutorials now. They are very good. Two things so far:

1) in the index to Landlab notebooks, Network Sediment Transporter is listed twice. Here is a screenshot:

2) I have walking through the NST tutorials using a Binder instance . Some of the links (specifically to other tutorials) point to a local port —  `localhost:8888` — and therefore don't function in my Binder instance.
