### 2020-07-13 #2341 @ebgoldstein
Ok, I have made it through the paper, docs, tutorials, and repo. The code works as described, and is well served by 4 tutorial notebooks. I think users will be able to understand the working of the code and how to modify it for their given use case. To my eye, this is a nice addition to Landlab,

I have one suggestion: The module described is in some sense a computational replication of Czuba 2018 — and i noticed that the tutorial for a real basin was a section of the the Methow River, which is the same as the original paper. I know this is JOSS and not Rescience C, but I think a mention of the results of this replication could be made. For example, the the authors could show (as a figure) or describe (in text) or put in the tutorials a statement about the comparison to Czuba 2018 Fig2 or Fig3... even a simple statement that 'the results are identical to the code of C18'.   I will leave this to @kthyng to decide if it is in scope.

also:
> Nothing will stop us from making all of the choices at once.
😆
