### 2020-07-15 ea73dfe
add link_with_node_status function
