### 2020-07-15 b4b30b0
Gt/space speed (#1215)
* wip space speed tuning

* remove test cy implementation

* clean up

* make pretty
