+++ landlab/bmi/bmi_bridge.py
@@ -15,6 +15,7 @@ exposes a Basic Modelling Interface.
 import inspect
 
 import numpy as np
+
 from bmipy import Bmi
 
 from ..core import load_params
