### 2020-07-15 b8c3ee2
remove usage of grid.links_by_tail_and_head_status
