### 2020-07-15 07e583f
link_with_node_status return an id array
