### 2020-07-15 ada6973
fix test for writing uint8 fields to netcdf
