### 2020-07-16 cb9ed5c
fix: unit tests for writing uint8 fields to netcdf (#1220)
* fix test for writing uint8 fields to netcdf

* require xarray >= 0.16
