### 2020-07-17 447be81
feat: utility for building matrices for finite difference/volume solutions (#1217)
* add properties for core-core, fv-core, and core-fv links

* add matrix creation utility 1st draft

* blackify

* add fns to __init__

* add tutorial notebook

* fix import in tutorial notebook

* make pretty

* lint

* add general link_by_tail_and_head_status method

* convert to list for doctest to avoid silly test fail around reporting or not the dtype of array

* update to use generic links_by_tail_and_head_status

* correct docstring

* lint

* add link_with_node_status function

* remove original in favor
[…]
