+++ landlab/components/tidal_flow/tidal_flow_calculator.py
@@ -1,0 +1,210 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Calculate cycle-averaged tidal flow field using approach of Mariotti (2018)
+"""
+
+import numpy as np
+from landlab import Component, RasterModelGrid, HexModelGrid
+from landlab.utils import make_core_node_matrix_var_coef
+from landlab.grid.mappers import map_mean_of_link_nodes_to_link
+
+_FOUR_THIRDS = 4.0 / 3.0
+_M2_PERIOD = (12.0 + (25.2 / 60.0)) * 3600.0  # M2 tidal period, in seconds
+
+
+class TidalFlowCalculator(Component):
+    """Calculate average flow over a tidal cycle."""
+
+    _name = "TidalFlowCalculator"
+
+    _unit_agnostic = False
+
+    _info = {
+        "flood_tide_flow__velocity": {
+            "dtype": float,
+            "intent": "out",
+            "optional": False,
+            "units": "m/s",
+            "mapping": "link",
+            "doc": "Horizontal flow velocity along links during flood tide",
+        },
+        "ebb_tide_flow__velocity": {
+            "dtype": float,
+            "intent": "out",
+            "optional": False,
+            "units": "m/s",
+            "mapping": "link",
+            "doc": "Horizontal flow velocity along links during ebb tide",
+        },
+        "topographic__elevation": {
+            "dtype": float,
+            "intent": "in",
+            "optional": False,
+            "units": "m",
+            "mapping": "node",
+            "doc": "Bathymetric/topographic elevation",
+        },
+        "mean_water__depth": {
+            "dtype": float,
+            "intent": "out",
+            "optional": False,
+            "units": "m",
+            "mapping": "node",
+            "doc": "Tidal mean water depth",
+        },
+    }
+
+    def __init__(
+        self,
+        grid,
+        tidal_range=1.0,
+        tidal_period=_M2_PERIOD,
+        roughness=0.01,
+        scale_velocity=1.0,
+        mean_sea_level=0.0,
+        min_water_depth=0.01,
+    ):
+        """Initialize TidalFlowCalculator."""
+
+        # Handle grid type
+        if isinstance(grid, RasterModelGrid):
+            self._grid_multiplier = 1.0
+        elif grid is HexModelGrid:
+            self._grid_multiplier = 1.5
+        else:
+            raise TypeError("Grid must be raster or hex.")
+
+        # Call base class methods to check existence of input fields,
+        # create output fields, etc.
+        super().__init__(grid)
+        self.initialize_output_fields()
+
+        # Get references to various fields, for convenience
+        self._elev = self.grid.at_node["topographic__elevation"]
+        self._water_depth = self.grid.at_node["mean_water__depth"]
+        self._flood_tide_vel = self.grid.at_link["flood_tide_flow__velocity"]
+        self._ebb_tide_vel = self.grid.at_link["ebb_tide_flow__velocity"]
+
+        # Handle inputs
+        # TODO: make properties
+        self.tidal_range = tidal_range
+        self._tidal_half_range = tidal_range / 2.0
+        self.tidal_period = tidal_period
+        self._tidal_half_period = tidal_period / 2.0
+        self._scale_velocity = scale_velocity
+        self.roughness = roughness  # TODO: replace with a field
+        self.mean_sea_level = mean_sea_level
+        self._min_depth = min_water_depth
+
+        # Make other data structures
+        self._water_depth_at_links = np.zeros(grid.number_of_links)
+        self._diffusion_coef_at_links = np.zeros(grid.number_of_links)
+
[…]
