### 2020-07-19 9322cad
add test for water shallower than tidal amplitude
