### 2020-07-20 #2341 @zsylvester
Yes, I will try to get to it this week. I just realized that I have not 'officially' accepted the invitation yet - when I click on the link (https://github.com/openjournals/joss-reviews/invitations) it says that it has expired.
