### 2020-07-21 #2341 @zsylvester
@kthyng It still does not let me to accept the invitation / go through the checklist.
