### 2020-07-22 #2341 @zsylvester
I have gone through the paper, the four Jupyter notebooks, scanned the repo, looked at the documentation, and ran the tests.  I think this is a nice addition to the Landlab package and it is likely to be of interest to anybody who is working on sediment transport problems in primarily non-depositional landscapes. The NetworkSedimentTransporter component is installed along with the Landlab package and I had no issues installing Landlab 2.0 (that happened a while ago). The three notebooks ran smoothly on Binder and they are easy to follow along, I can see how the second one could form a starting point for playing with someone's own dataset. I have run the fourth one locally.

Here are a few comments / suggestions:
- I agree with @ebgoldstein that a statement about how the model compares with Czuba (2018) would be useful.
- Along the same lines, it would be nice to see a bit larger / more complex model in the notebook with the shapefile example, maybe something similar to what is shown in Czuba (2018), and show how it could be used to provide insights into some aspects of sediment transport.
- The profiling and scaling notebook could use a bit more explanation / documentation.
- Many of the links in the notebooks do not work.
- Link to Jupyter notebook at the end of the manuscript does not work.
- I think 'MATLAB' is the correct usage, not 'Matlab'.
- 'computational infrastructure built in order to create the NetworkSedimentTransporter' instead of  'computational infrastructure created in order to create the NetworkSedimentTransporter'.
- '...required the addition of two new data structures' instead of '...required the creation of two new data structures'.
- 'Landlab' instead of 'landlab' in the Hobley et al. reference.
