### 2020-07-23 61060bc
add test with hex grid
