+++ landlab/components/tidal_flow/test_tidal_flow_calculator.py
@@ -8,7 +8,7 @@ Created on Sun Jul 12 10:22:59 2020
 
 from tidal_flow_calculator import TidalFlowCalculator
 from numpy.testing import assert_array_almost_equal
-from landlab import RasterModelGrid
+from landlab import RasterModelGrid, HexModelGrid
 
 
 def test_constant_depth_deeper_than_tidal_amplitude():
@@ -24,7 +24,7 @@ def test_constant_depth_deeper_than_tidal_amplitude():
 
     velocity coefficient, Cv = h^(4/3) / n^2 chi = 1.6e5 m/s
     diffusion coefficient, D = h^(7/3) / n^2 chi = 1.28e6 m2/s
-    inundation rate (I) = r / T = 5e-5 m/s
+    inundation rate (I) = 2 r / T = 5e-5 m/s
 
     Domain, L: 300 m, 1D
 
@@ -45,22 +45,26 @@ def test_constant_depth_deeper_than_tidal_amplitude():
         0.000625, 0.00125, 0.001875
     """
     grid = RasterModelGrid((3, 5), xy_spacing=100.0)
-    z = grid.add_zeros('topographic__elevation', at='node')
+    z = grid.add_zeros("topographic__elevation", at="node")
     z[:] = -8.0
     grid.set_closed_boundaries_at_grid_edges(False, True, True, True)
 
     tfc = TidalFlowCalculator(grid, tidal_period=4.0e4)
     tfc.run_one_step()
 
-    assert_array_almost_equal(grid.at_link["ebb_tide_flow__velocity"][10:13],
-                              [0.000625, 0.00125, 0.001875])
-    assert_array_almost_equal(grid.at_link["flood_tide_flow__velocity"][10:13],
-                              [-0.000625, -0.00125, -0.001875])
+    assert_array_almost_equal(
+        grid.at_link["ebb_tide_flow__velocity"][10:13], [0.000625, 0.00125, 0.001875]
+    )
+    assert_array_almost_equal(
+        grid.at_link["flood_tide_flow__velocity"][10:13],
+        [-0.000625, -0.00125, -0.001875],
+    )
+
 
 def test_constant_depth_deeper_than_tidal_amplitude_alt_grid():
     """Test velocity calculation with different grid orientation."""
     grid = RasterModelGrid((5, 3), xy_spacing=100.0)
-    z = grid.add_zeros('topographic__elevation', at='node')
+    z = grid.add_zeros("topographic__elevation", at="node")
     z[:] = -8.0
     grid.set_closed_boundaries_at_grid_edges(True, False, True, True)
 
@@ -69,11 +73,10 @@ def test_constant_depth_deeper_than_tidal_amplitude_alt_grid():
 
     links_to_test = [8, 13, 18]
     ebb_vel = grid.at_link["ebb_tide_flow__velocity"][links_to_test]
-    assert_array_almost_equal(ebb_vel,
-                              [0.000625, 0.00125, 0.001875])
+    assert_array_almost_equal(ebb_vel, [0.000625, 0.00125, 0.001875])
     flood_vel = grid.at_link["flood_tide_flow__velocity"][links_to_test]
-    assert_array_almost_equal(flood_vel,
-                              [-0.000625, -0.00125, -0.001875])
+    assert_array_almost_equal(flood_vel, [-0.000625, -0.00125, -0.001875])
+
 
 def test_constant_depth_shallower_than_tidal_amplitude():
     r"""Test velocity calculation under the following conditions:
@@ -88,14 +91,14 @@ def test_constant_depth_shallower_than_tidal_amplitude():
 
     velocity coefficient, Cv = h^(4/3) / n^2 chi = 1.6e5 m/s
[…]
