### 2020-07-24 f3b4435
make roughness a scalar or array, and other params properties
