### 2020-07-24 5a58afa
add option for sparse matrix
