### 2020-07-25 ed959c7
add tests for 100% coverage
