### 2020-07-25 3f322a2
move test file to tests folder
