+++ landlab/components/groundwater/dupuit_percolator.py
@@ -328,12 +328,13 @@ class GroundwaterDupuitPercolator(Component):
             This parameter is only used with ``run_with_adaptive_time_step_solver``
             and must be greater than zero.
             Default = 0.8
-        callback_fun: function(grid, substep_dt, \*\*kwargs)
+        callback_fun: function(grid, recharge_rate, substep_dt, \*\*kwargs)
             Optional function that will be executed at the end of each sub-timestep
             in the run_with_adaptive_time_step_solver method. Intended purpose
             is to write output not otherwise visible outside of the method call.
             The function should have two required arguments:
             grid: the ModelGrid instance used by GroundwaterDupuitPercolator
+            recharge_rate: an array at node that is the component recharge rate
             substep_dt: the length of the current substep determined internally
             by run_with_adaptive_time_step_solver to meet stability criteria.
         """
@@ -752,6 +753,6 @@ class GroundwaterDupuitPercolator(Component):
             remaining_time -= substep_dt
             self._num_substeps += 1
 
-            self._callback_fun(self._grid, substep_dt)
+            self._callback_fun(self._grid, self.recharge, substep_dt)
 
         self._qsavg[:] = qs_cumulative / dt
