### 2020-07-29 e5bdd69
Initial edits in response to reviews
Note: tests for new active layer thickness knob still need to be completed..
