+++ joss/in_preparation/network_sediment_transporter/paper.md
@@ -43,15 +43,15 @@ Coarse sediment (sand, gravel, and cobbles) moves downstream through river netwo
 
 Most computational morphodynamic models take an Eulerian approach, which tracks changes in bed elevation through time as a function of the spatial gradient in sediment flux [e.g., @ParkerEbook]. These models directly compute bed elevation change and sediment flux throughout the domain. One of the major drawbacks with Eulerian morphodynamic models is the difficulty in being able to 'tag' individual sediment particles to answer questions about how an individual sediment particle/input may move, when it might arrive, and what affect it will have on river morphology when it arrives downstream. To overcome this drawback and to more easily extend morphodynamic models to entire river networks, recent work has focused on developing river-network based Lagrangian sediment transport models, which track the locations of individual sediment units on a river network.  
 
-A more comprehensive overview of river-network based sediment transport models is described by @Czubaetal2017. Of most relevance to the work described herein, @Czuba2018 introduced a network-based, Lagrangian bed material morphodynamic model that tracks the motion of individual units (referred to as “parcels”) of sediment through a river network. This model has been applied to post-wildfire debris-flow sediment movement through a river network in Utah [@Murphyetal2019]. This approach improves on the existing morphodynamic models by: (1) accounting for the full river network, rather than a single longitudinal profile, (2) allowing the user to ‘tag’ particular sediment inputs and track their fate through time. Despite its advances, this existing network sediment transport model, however, has two notable drawbacks: 1) it written in a proprietary scripting language (MatLab), and 2) it is not explicitly designed to be interoperable with other Earth-surface models, such as streamflow or landslide models.
+A more comprehensive overview of river-network based sediment transport models is described by @Czubaetal2017. Of most relevance to the work described herein, @Czuba2018 introduced a network-based, Lagrangian bed material morphodynamic model that tracks the motion of individual units (referred to as “parcels”) of sediment through a river network. This model has been applied to post-wildfire debris-flow sediment movement through a river network in Utah [@Murphyetal2019]. This approach improves on the existing morphodynamic models by: (1) accounting for the full river network, rather than a single longitudinal profile, (2) allowing the user to ‘tag’ particular sediment inputs and track their fate through time. Despite its advances, this existing network sediment transport model, however, has two notable drawbacks: 1) it written in a proprietary scripting language (MATLAB), and 2) it is not explicitly designed to be interoperable with other Earth-surface models, such as streamflow or landslide models.
 
-Here, we present software that overcomes these two drawbacks, translating the network sediment transport model of @Czuba2018 into Landlab, a modular, Python-based package for the modeling of Earth-surface dynamics. Landlab is an Open Source Python package for modeling Earth-surface processes [@Hobley2017Creative; @Barnhart2020Short]. It was designed as a modular framework, hosting a variety of process components such as flow routing, hillslope diffusion, and stream power erosion that function on a common set of landscape model grids. The ``NetworkSedimentTransporter`` is the newest of these components. We first describe computational infrastructure created in order to create the ``NetworkSedimentTransporter`` and then describe the new component itself.
[…]
