+++ landlab/components/network_sediment_transporter/network_sediment_transporter.py
@@ -7,7 +7,7 @@ in bed material grain size and river bed elevation. Model framework
 described in Czuba (2018). Additions include: particle abrasion, variable
 active layer thickness (Wong et al., 2007).
 
-.. codeauthor:: Allison Pfeiffer, Katy Barnhart, Jon Czuba
+.. codeauthor:: Allison Pfeiffer, Katy Barnhart, Jon Czuba, Eric Hutton
 
 Created on Tu May 8, 2018
 Last edit was sometime after February 2020
@@ -25,6 +25,7 @@ from landlab.data_record import DataRecord
 from landlab.grid.network import NetworkModelGrid
 
 _SUPPORTED_TRANSPORT_METHODS = ["WilcockCrowe"]
+_SUPPORTED_ACTIVE_LAYER_METHODS = ["WongParker", "GrainSizeDependent", "Constant10cm"]
 
 _REQUIRED_PARCEL_ATTRIBUTES = [
     "time_arrival_in_link",
@@ -163,6 +164,7 @@ class NetworkSedimentTransporter(Component):
     ...         g=9.81,
     ...         fluid_density=1000,
     ...         transport_method="WilcockCrowe",
+    ...         active_layer_method="WongParker"
     ...     )
 
     >>> dt = 60  # (seconds) 1 min timestep
@@ -261,6 +263,9 @@ class NetworkSedimentTransporter(Component):
         g=scipy.constants.g,
         fluid_density=1000.0,
         transport_method="WilcockCrowe",
+        active_layer_method="WongParker",
+        active_layer_d_multiplier=2
+
     ):
         """
         Parameters
@@ -294,6 +299,9 @@ class NetworkSedimentTransporter(Component):
         transport_method: string
             Sediment transport equation option. Default (and currently only)
             option is "WilcockCrowe".
+        active_layer_method: string, optional
+            Option for treating sediment active layer as a constant or variable
+            (default, "WongParker")
         """
         if not isinstance(grid, NetworkModelGrid):
             msg = "NetworkSedimentTransporter: grid must be NetworkModelGrid"
@@ -365,6 +373,15 @@ class NetworkSedimentTransporter(Component):
         # for the transport method.
         if self._transport_method == "WilcockCrowe":
             self._update_transport_time = self._calc_transport_wilcock_crowe
+
+        if active_layer_method in _SUPPORTED_ACTIVE_LAYER_METHODS:
+            self._active_layer_method = active_layer_method
+        else:
+            msg = "NetworkSedimentTransporter: Active layer method not supported."
+            raise ValueError(msg)
+
+        if self._active_layer_method == "GrainSizeDependent":
+            self._active_layer_d_multiplier = active_layer_d_multiplier
 
         # save reference to key fields
         self._width = self._grid.at_link["channel_width"]
@@ -498,33 +515,44 @@ class NetworkSedimentTransporter(Component):
             fill_value=0.0,
         )
 
-        # Wong et al. (2007) approximation for active layer thickness.
-        # NOTE: calculated using grain size and grain density calculated for
-        # the active layer grains in each link at the **previous** timestep.
-        # This circumvents the need for an iterative scheme to determine grain
-        # size of the active layer before determining which grains are in the
-        # active layer.
+        if self._active_layer_method == "WongParker":
+            # Wong et al. (2007) approximation for active layer thickness.
+            # NOTE: calculated using grain size and grain density calculated for
+            # the active layer grains in each link at the **previous** timestep.
+            # This circumvents the need for an iterative scheme to determine grain
+            # size of the active layer before determining which grains are in the
+            # active layer.
 
-        # calculate tau
-        tau = (
-            self._fluid_density
-            * self._g
-            * self._grid.at_link["channel_slope"]
[…]
