### 2020-07-30 21ca4f4
cythonize part of make_core_node_matrix
