+++ landlab/utils/cfuncs.pyx
@@ -1,0 +1,47 @@
+import numpy as np
+cimport numpy as np
+cimport cython
+
+DTYPE_INT = np.int
+ctypedef np.int_t DTYPE_INT_t
+
+DTYPE_FLOAT = np.double
+ctypedef np.double_t DTYPE_FLOAT_t
+
+
+def fill_matrix(np.ndarray[DTYPE_INT_t, ndim=1] core2core,
+                np.ndarray[DTYPE_INT_t, ndim=1] core2fv,
+                np.ndarray[DTYPE_INT_t, ndim=1] fv2core,
+                np.ndarray[DTYPE_INT_t, ndim=1] node_at_link_tail,
+                np.ndarray[DTYPE_INT_t, ndim=1] node_at_link_head,
+                np.ndarray[DTYPE_INT_t, ndim=1] matrow,
+                np.ndarray[DTYPE_FLOAT_t, ndim=1] value,
+                np.ndarray[DTYPE_FLOAT_t, ndim=2] rhs,
+                object mat,
+                ):
+
+    cdef int ln, t, h
+
+    for ln in core2core:
+        t = node_at_link_tail[ln]
+        h = node_at_link_head[ln]
+        mat[matrow[t], matrow[t]] -= 1.0
+        mat[matrow[h], matrow[h]] -= 1.0
+        mat[matrow[t], matrow[h]] = 1.0
+        mat[matrow[h], matrow[t]] = 1.0
+
+    # Handle core-to-fv links
+    for ln in core2fv:
+        t = node_at_link_tail[ln]
+        h = node_at_link_head[ln]
+        mat[matrow[t], matrow[t]] -= 1
+        rhs[matrow[t]] -= value[h]
+
+    # Handle fv-to-core links
+    for ln in fv2core:
+        t = node_at_link_tail[ln]
+        h = node_at_link_head[ln]
+        mat[matrow[h], matrow[h]] -= 1
+        rhs[matrow[h]] -= value[t]
+
+    return mat, rhs
