### 2020-07-31 819e6a0
cythonize part of make_core_node_matrix_var_coef
