### 2020-07-31 a4d43c4
revisions to clarify replication/difference with C2018
