### 2020-08-01 938fdb9
test alternative cython implementations
