### 2020-08-01 2a99014
use correct method for calculating effective water depth
