### 2020-08-01 1b9d733
use correct eqn for tidal-average water depth, and update tests
