### 2020-08-03 7868f58
Corrected path to func
Corrected path to funcs in spatial_disturbance.py
