### 2020-08-03 0939a3d
Fix test function fields
