### 2020-08-03 a4dbed6
add mapper for link vectors to links, and specialized implementation for hex
