+++ landlab/grid/mappers.py
@@ -22,6 +22,7 @@ Grid mapping functions
     ~landlab.grid.mappers.map_downwind_node_link_mean_to_node
     ~landlab.grid.mappers.map_value_at_upwind_node_link_max_to_node
     ~landlab.grid.mappers.map_value_at_downwind_node_link_max_to_node
+    ~landlab.grid.mappers.map_link_vector_components_to_node
     ~landlab.grid.mappers.dummy_func_to_demonstrate_docstring_modification
 
 Each link has a *tail* and *head* node. The *tail* nodes are located at the
@@ -1453,6 +1454,29 @@ def map_link_vector_sum_to_patch(grid, var_name, ignore_inactive_links=True, out
         vert_sum = np.sum(vert_vals_at_patches, axis=1, out=out[1])
 
     return out
+
+
+def map_link_vector_components_to_node(grid, data_at_link):
+    """Map (x,y) components of link data data_at_link onto nodes.
+
+    Examples
+    --------
+    >>> from landlab import HexModelGrid
+    >>> grid = HexModelGrid((3, 3))
+    >>> link_data = np.zeros(grid.number_of_links) + 0.5 * 3.0**0.5
+    >>> link_data[grid.link_with_angle(0.0)] = 0.0
+    >>> vx, vy = map_link_vector_components_to_node(grid, link_data)
+    >>> vy
+    array([ 0.,  0.,  0.,  0.,  1.,  1.,  0.,  0.,  0.,  0.])
+    """
+    from landlab import HexModelGrid
+
+    if isinstance(grid, HexModelGrid):
+        from .hex_mappers import map_link_vector_components_to_node_hex
+
+        return map_link_vector_components_to_node_hex(grid, data_at_link)
+    else:
+        raise NotImplementedError("Only available for HexModelGrid")
 
 
 def dummy_func_to_demonstrate_docstring_modification(grid, some_arg):
