+++ landlab/grid/hex_mappers.py
@@ -1,0 +1,140 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""Grid element mappers that are specific to hex grids.
+
+Mapping functions unique to hex grids
++++++++++++++++++++++++++++++++++++++
+
+.. autosummary::
+
+    ~landlab.grid.hex_mappers.map_link_vector_components_to_node_hex
+"""
+
+import numpy as np
+
+LINK_TO_EAST = 0  # array column of link to east/right of node, horizontal grid
+LINK_TO_NNE = 1  # array column of link to nne/upper right of node, horizontal grid
+LINK_TO_NNW = 2  # array column of link to nnw/upper left of node, horizontal grid
+LINK_TO_WEST = 3  # array column of link to west/left of node, horizontal grid
+LINK_TO_SSW = 4  # array column of link to ssw/lower left of node, horizontal grid
+LINK_TO_SSE = 5  # array column of link to sse/lower right of node, horizontal grid
+
+LINK_TO_ENE = 0  # array column of link to ene/upper right of node, vertical grid
+LINK_TO_NORTH = 1  # array column of link to north/top of node, vertical grid
+LINK_TO_WNW = 2  # array column of link to wnw/upper left of node, vertical grid
+LINK_TO_WSW = 3  # array column of link to wsw/lower left of node, vertical grid
+LINK_TO_SOUTH = 4  # array column of link to south/bottom of node, vertical grid
+LINK_TO_ESE = 5  # array column of link to ese/lower right of node, vertical grid
+
+SIN60 = np.sin(np.deg2rad(60.0))
+
+
+def map_link_vector_components_to_node_hex(grid, data_at_link):
+    """Map (x,y) components of link data data_at_link onto nodes of hex grid.
+
+    Parameters
+    ----------
+    grid : HexModelGrid
+        Landlab HexModelGrid object
+    data_at_links : ndarray of float x number of links
+        Data to be mapped
+
+    Examples
+    --------
+    >>> from landlab import HexModelGrid
+    >>> from landlab.grid.mappers import map_link_vector_components_to_node
+    >>> grid = HexModelGrid((3, 3))
+    >>> link_data = np.zeros(grid.number_of_links) + 0.5 * 3.0**0.5
+    >>> link_data[grid.link_with_angle(0.0)] = 0.0
+    >>> vx, vy = map_link_vector_components_to_node(grid, link_data)
+    >>> vx
+    array([ 0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.])
+    >>> vy
+    array([ 0.,  0.,  0.,  0.,  1.,  1.,  0.,  0.,  0.,  0.])
+    >>> link_data = np.arange(grid.number_of_links)
+    >>> vx, vy = map_link_vector_components_to_node(grid, link_data)
+    >>> vx
+    array([ 0. ,  0. ,  0. ,  0. ,  8.5,  9.5,  0. ,  0. ,  0. ,  0. ])
+    >>> link_data = np.zeros(grid.number_of_links) + 0.5 * 3.0**0.5
+    >>> link_data[grid.link_with_angle(120.0)] = 0.0
+    >>> vx, vy = map_link_vector_components_to_node(grid, link_data)
+    >>> np.round(vx, 3)
+    array([ 0. ,  0. ,  0. ,  0. ,  0.866,  0.866,  0. ,  0. ,  0. ,  0. ])
+    >>> vy
+    array([ 0. ,  0. ,  0. ,  0. ,  0.5,  0.5,  0. ,  0. ,  0. ,  0. ])
+
+    >>> grid = HexModelGrid((3, 3), orientation='vertical')
[…]
