### 2020-08-03 1adff84
add raster version of link vector mapper
