+++ landlab/grid/mappers.py
@@ -1461,7 +1461,13 @@ def map_link_vector_components_to_node(grid, data_at_link):
 
     Examples
     --------
-    >>> from landlab import HexModelGrid
+    >>> from landlab import RasterModelGrid, HexModelGrid
+    >>> grid = RasterModelGrid((3, 4))
+    >>> link_data = np.arange(grid.number_of_links)
+    >>> vx, vy = map_link_vector_components_to_node(grid, link_data)
+    >>> vx[5:7]
+    array([ 7.5, 8.5])
+
     >>> grid = HexModelGrid((3, 3))
     >>> link_data = np.zeros(grid.number_of_links) + 0.5 * 3.0**0.5
     >>> link_data[grid.link_with_angle(0.0)] = 0.0
@@ -1469,12 +1475,16 @@ def map_link_vector_components_to_node(grid, data_at_link):
     >>> vy
     array([ 0.,  0.,  0.,  0.,  1.,  1.,  0.,  0.,  0.,  0.])
     """
-    from landlab import HexModelGrid
+    from landlab import HexModelGrid, RasterModelGrid
 
     if isinstance(grid, HexModelGrid):
         from .hex_mappers import map_link_vector_components_to_node_hex
 
         return map_link_vector_components_to_node_hex(grid, data_at_link)
+    elif isinstance(grid, RasterModelGrid):
+        from .raster_mappers import map_link_vector_components_to_node_raster
+
+        return map_link_vector_components_to_node_raster(grid, data_at_link)
     else:
         raise NotImplementedError("Only available for HexModelGrid")
 
