### 2020-08-04 88773f9
Update test_active_layer_methods.py
Test grain size dependent and constant active layer options.
