+++ joss/in_preparation/network_sediment_transporter/paper.md
@@ -45,19 +45,19 @@ Most computational morphodynamic models take an Eulerian approach, which tracks 
 
 A more comprehensive overview of river-network based sediment transport models is described by @Czubaetal2017. Of most relevance to the work described herein, @Czuba2018 introduced a network-based, Lagrangian bed material morphodynamic model that tracks the motion of individual units (referred to as “parcels”) of sediment through a river network. The model presented by @Czuba2018 has been applied to post-wildfire debris-flow sediment movement through a river network in Utah [@Murphyetal2019]. Czuba's approach improves on the existing morphodynamic models by: (1) accounting for the full river network, rather than a single longitudinal profile, (2) allowing the user to ‘tag’ particular sediment inputs and track their fate through time. Despite its advances, this existing network sediment transport model, however, has two notable drawbacks: 1) it written in a proprietary scripting language (MATLAB), and 2) it is not explicitly designed to be interoperable with other Earth-surface models, such as streamflow or landslide models.
 
-Here, we present software that overcomes these two drawbacks, translating and expanding upon the network sediment transport model of @Czuba2018 into Landlab, a modular, Python-based package for the modeling of Earth-surface dynamics. Landlab is an Open Source Python package for modeling Earth-surface processes [@Hobley2017Creative; @Barnhart2020Short]. It was designed as a modular framework, hosting a variety of process components such as flow routing, hillslope diffusion, and stream power erosion that function on a common set of landscape model grids. The ``NetworkSedimentTransporter`` is the newest of these components. We first describe computational infrastructure built in order to create the ``NetworkSedimentTransporter`` and then describe the new component itself.
+Here, we present software that overcomes these two drawbacks, translating and expanding upon the network sediment transport model of @Czuba2018 in Landlab, a modular, Python-based package for the modeling of Earth-surface dynamics. Landlab is an Open Source Python package for modeling Earth-surface processes [@Hobley2017Creative; @Barnhart2020Short]. It was designed as a modular framework, hosting a variety of process components such as flow routing, hillslope diffusion, and stream power erosion that function on a common set of landscape model grids. The ``NetworkSedimentTransporter`` is the newest of these components. We first describe computational infrastructure built in order to create the ``NetworkSedimentTransporter`` and then describe the new component itself.
 
 The creation of the ``NetworkSedimentTransporter`` required the addition of two new data structures in the Landlab framework. First, the ``NetworkModelGrid``, which represents the model domain as connected nodes and links. Second, the ``DataRecord``, which stores a generic set of items in time and on the model grid. It is used here to store all attributes associated with the sediment parcels that move into, through, and out of the model domain.
 
 In the ``NetworkSedimentTransporter``, sediment is represented as "parcels"-a quantity of sediment grains with common attributes such as grain diameter, lithology, and density. Each parcel is transported, buried, and eroded as a coherent unit. The river network is represented as a series of links and nodes on a ``NetworkModelGrid``. Each time the ``NetworkSedimentTransporter`` is run forward in time, the set of parcels that are in active transport is identified based on the flow conditions and bed surface grain size in each link, transport distances are calculated for all active parcels based on the @WilcockCrowe2003 equations, and parcels move through links on the network by updating their locations based on their transport distances [@Czuba2018]. As a result of parcel redistribution, the elevation of nodes and slope of the links evolves [@Czubaetal2017; @Czuba2018].
 
[…]
