### 2020-08-07 60a2ab2
Changed "submodule" to "component" in text
As per email discussion
