+++ landlab/components/groundwater/dupuit_percolator.py
@@ -406,18 +406,24 @@ class GroundwaterDupuitPercolator(Component):
 
     @callback_fun.setter
     def callback_fun(self, new_val):
-        try: # New style callback function.
+        try:  # New style callback function.
             new_val(self._grid, self.recharge, 0.0, **self.callback_kwds)
             self._old_style_callback = False
             self._callback_fun = new_val
         except TypeError:
-            try: # Old style callback function, will work, but warn.
+            try:  # Old style callback function, will work, but warn.
                 new_val(self._grid, 0.0, **self.callback_kwds)
                 self._callback_fun = new_val
                 self._old_style_callback = True
-                warn('Callback functions with two arguments (grid, substep_dt) are deprecated and will be removed in future versions', DeprecationWarning)
-            except TypeError as error: # Nonfunctional callback function.
-                raise ValueError('%s. Please supply a callback function with the form function(grid, recharge_rate, substep_dt, \*\*kwargs)'%error)
+                warn(
+                    "Callback functions with two arguments (grid, substep_dt) are deprecated and will be removed in future versions",
+                    DeprecationWarning,
+                )
+            except TypeError as error:  # Nonfunctional callback function.
+                raise ValueError(
+                    "%s. Please supply a callback function with the form function(grid, recharge_rate, substep_dt, \*\*kwargs)"
+                    % error
+                )
 
     @property
     def courant_coefficient(self):
@@ -794,6 +800,8 @@ class GroundwaterDupuitPercolator(Component):
             if self._old_style_callback:
                 self._callback_fun(self._grid, substep_dt, **self.callback_kwds)
             else:
-                self._callback_fun(self._grid, self.recharge, substep_dt, **self.callback_kwds)
+                self._callback_fun(
+                    self._grid, self.recharge, substep_dt, **self.callback_kwds
+                )
 
         self._qsavg[:] = qs_cumulative / dt
