### 2020-08-10 9f3824e
Apply suggestions from Katy's code review
Co-authored-by: Katy Barnhart <krbarnhart@usgs.gov>
