### 2020-08-11 #2341 @pfeiffea
@kthyng Changes are nearly complete-- we're just waiting for them to be published in the newest release of landlab/landlab. We should have a new draft to you all in the next few days.
