### 2020-08-17 695269a
release: v2.2.0 (#1234)
* remove fetched folder.

* Add HeightAboveDrainageCalculator component (#1207)

* Begin HeightAboveDrainage component

* Add warn

* initialize fields, fix indentation

* delint

* Start hand tests

* Add description, change pit handling

* change name and docstring

* Finish update name

* Fix docstring tests?

* Use stack

* delint

* Fix optional field

* fix components init and add parameter description

* Channel mask info

* fix channel mask in init

* delint

* Delint tests

* little stuff

* Gt/make w0 public in weatherer (#1196)

* update import syntax for use
[…]
