+++ landlab/components/flow_accum/flow_accumulator.py
@@ -723,12 +723,7 @@ class FlowAccumulator(Component):
         self._node_cell_area = node_cell_area
 
         # STEP 2:
-        # identify Flow Director method, save name, import and initialize the correct
-        # flow director component if necessary
-        self._add_director(flow_director)
-        self._add_depression_finder(depression_finder)
-
-        # This component will track of the following variables.
+        # This component will track the following variables.
         # Attempt to create each, if they already exist, assign the existing
         # version to the local copy.
 
@@ -751,6 +746,13 @@ class FlowAccumulator(Component):
 
         self._D_structure = self._grid.BAD_INDEX * grid.ones(at="link", dtype=int)
         self._nodes_not_in_stack = True
+
+        # STEP 3:
+        # identify Flow Director method, save name, import and initialize the
+        # correct flow director component if necessary; same with
+        # lake/depression handler, if specified.
+        self._add_director(flow_director)
+        self._add_depression_finder(depression_finder)
 
         if len(self._kwargs) > 0:
             kwdstr = " ".join(list(self._kwargs.keys()))
@@ -957,7 +959,7 @@ class FlowAccumulator(Component):
 
     def _add_depression_finder(self, depression_finder):
         """Test and add the depression finder component."""
-        PERMITTED_DEPRESSION_FINDERS = ["DepressionFinderAndRouter"]
+        PERMITTED_DEPRESSION_FINDERS = ["DepressionFinderAndRouter", "LakeMapperBarnes"]
 
         # now do a similar thing for the depression finder.
         self._depression_finder_provided = depression_finder
@@ -965,7 +967,19 @@ class FlowAccumulator(Component):
 
             # collect potential kwargs to pass to depression_finder
             # instantiation
-            potential_kwargs = ["routing"]
+            potential_kwargs = [
+                "routing",
+                "pits",
+                "reroute_flow",
+                "surface",
+                "method",
+                "fill_flat",
+                "fill_surface",
+                "redirect_flow_steepest_descent",
+                "reaccumulate_flow",
+                "ignore_overfill",
+                "track_lakes",
+            ]
             kw = {}
             for p_k in potential_kwargs:
                 if p_k in self._kwargs.keys():
@@ -1006,10 +1020,14 @@ class FlowAccumulator(Component):
             # depression finder is provided as a string.
             if isinstance(self._depression_finder_provided, str):
 
-                from landlab.components import DepressionFinderAndRouter
+                from landlab.components import (
+                    DepressionFinderAndRouter,
+                    LakeMapperBarnes,
+                )
 
                 DEPRESSION_METHODS = {
-                    "DepressionFinderAndRouter": DepressionFinderAndRouter
+                    "DepressionFinderAndRouter": DepressionFinderAndRouter,
+                    "LakeMapperBarnes": LakeMapperBarnes,
                 }
 
                 try:
@@ -1067,6 +1085,9 @@ class FlowAccumulator(Component):
         else:
             self._depression_finder = None
 
+    def pits_present(self):
+        return np.any(self._grid.at_node["flow__sink_flag"][self._grid.core_nodes])
+
     def accumulate_flow(self, update_flow_director=True, update_depression_finder=True):
         """Function to make FlowAccumulator calculate drainage area and
         discharge.
@@ -1122,8 +1143,8 @@ class FlowAccumulator(Component):
             # At the moment, no depression finders work with to-many, so it
             # lives here
             if self._depression_finder_provided is not None:
-                if update_depression_finder:
-                    self._depression_finder.map_depressions()
[…]
