### 2020-08-17 #2341 @pfeiffea
@kthyng, @ebgoldstein and @zsylvester:

Thank you for your thorough reading and suggestions on the first draft of our submission. We have edited the component, notebooks, and text. The updated version is now published on the main branch of Landlab (v2.2.0): https://github.com/landlab/landlab

Our response to individual comments are included below.

**A)** Computational Replication / Comparison to Czuba (2018)

Both reviewers bring up making a clearer comparison between this paper and the work presented in [Czuba (2018)](https://www.sciencedirect.com/science/article/abs/pii/S0169555X18303349?via%3Dihub). Specifically the following two comments:

From @ebgoldstein

> I have one suggestion: The module described is in some sense a computational replication of Czuba 2018 — and i noticed that the tutorial for a real basin was a section of the the Methow River, which is the same as the original paper. I know this is JOSS and not Rescience C, but I think a mention of the results of this replication could be made. For example, the the authors could show (as a figure) or describe (in text) or put in the tutorials a statement about the comparison to Czuba 2018 Fig2 or Fig3... even a simple statement that 'the results are identical to the code of C18'. I will leave this to @kthyng to decide if it is in scope.

From @zsylvester:

> I agree with @ebgoldstein that a statement about how the model compares with Czuba (2018) would be useful.

**Response**:
The NetworkSedimentTransporter is not a direct translation of the model presented in Czuba (2018) and thus we don't think an attempt at replicating the results of this paper makes sense. However, we interpret these comments to indicate two important points: (a) that our original text does not make clear that this code is not a direct translation of Czuba (2018) and (b) the specific differences between this implementation and that of Czuba (2018) are not sufficiently clear.

We felt that the best way to address this was to add the following text:

> Our implementation is not a direct translation of the model implemented in MATLAB and described in @Czuba2018. In addition, we add two new elements to the model: variable sediment density and bed-material abrasion, and specification of active layer thickness.

We also revised the text describing the differences for clarity.

**B)** Profiling and Scaling Notebook

@zsylvester pointed out the bare-bones nature of the profiling and scaling notebook:

> The profiling and scaling notebook could use a bit more explanation / documentation.

Thank you for pointing out this oversight. We have greatly expanded this notebook. It should be far more helpful to users in its revised form.

**C)** A More Complex Model Notebook

@zsylvester made a request for a notebook demonstrating a more sophisticated use case:

>[It] would be nice to see a bit larger / more complex model in the notebook with the shapefile example, maybe something similar to what is shown in Czuba (2018), and show how it could be used to provide insights into some aspects of sediment transport.

We feel that a more complex model would be out of scope for a Landlab tutorial. The goal of the Landlab tutorials is to demonstrate a simple, generalized use case of each component, rather than to explore specific scientific questions. Previous work on network-scale Lagrangian sediment transport has demonstrated the value of these models to our understanding of basin scale sediment dynamics (see [Gran and Czuba, 2017](http://geomorphology.sese.asu.edu/Papers/Gran_2017_sediment_pulse.pdf) , [Czuba and Foufoula-Georgiou, 2015](https://agupubs.onlinelibrary.wiley.com/doi/full/10.1002/2014WR016139) ). Furthermore, the updated profiling and scaling notebook will provide potential users a sense for the efficiency of this component for larger models.

**D)** New Active Layer Options

The original version of the paper text referenced multiple methods for determining active layer thickness:

> Many sediment transport models [e.g., @Cui2007b; @Czuba2018] represent the mobile portion of the grains on the riverbed at any given time as an "active layer" of constant thickness. All grains in this layer are transported, whereas all grains below this layer are immobile. Within ``NetworkSedimentTransporter`` the user has the option to specify active layer thickness as a constant value or a multiple of the mean grain size in each link. Alternatively, we incorporated the formulation of @Wongetal2007 to calculate an active layer thickness for each link in the network at each timestep as a function of Shields stress and median grain diameter.

We realized that this functionality was not included in the version of the code that was originally merged into Landlab and submitted to JOSS. We have included this ‘knob’ (and associated tests) in this round of edits.

**E)** Term used to describe NST

After a discussion between authors, we have decided to call the ``NetworkSedimentTransporter`` a “component” instead of “submodule”. The difference is semantic, though we think that “component” may be clearer for readers.

**F)** Minor Edits

We have fixed the typos, duplicate notebook listings, and broken links pointed out by @ebgoldstein and @zsylvester. Thank you for finding these!
