### 2020-08-17 ddc4c8f
add faster get_core_node_matrix
