+++ landlab/utils/__init__.py
@@ -2,10 +2,8 @@
 
 from .add_halo import add_halo
 
-# import landlab.utils.count_repeats
-# from landlab.utils.count_repeats import count_repeats
 from .count_repeats import count_repeated_values
-from .matrix import make_core_node_matrix, make_core_node_matrix_var_coef
+from .matrix import get_core_node_at_node, get_core_node_matrix
 from .return_array import return_array_at_link, return_array_at_node
 from .source_tracking_algorithm import (
     convert_arc_flow_directions_to_landlab_node_ids,
@@ -35,6 +33,6 @@ __all__ = [
     "StablePriorityQueue",
     "return_array_at_node",
     "return_array_at_link",
-    "make_core_node_matrix",
-    "make_core_node_matrix_var_coef",
+    "get_core_node_at_node",
+    "get_core_node_matrix",
 ]
