+++ landlab/utils/matrix.py
@@ -3,63 +3,70 @@
 """Functions to set up a finite-volume solution matrix for a landlab grid."""
 
 import numpy as np
-from scipy.sparse import lil_matrix
-from .cfuncs import fill_matrix, fill_matrix_with_coefficients
+from scipy.sparse import csc_matrix
+
+from ._matrix import fill_right_hand_side, get_matrix_diagonal_elements, get_matrix_diagonal_elements_with_coef
 
 
-def matrix_row_at_node(grid):
-    """Create and return a node array with corresponding matrix row, or -1 if none.
+def get_core_node_at_node(grid):
+    """Get node ids as numbered by core nodes.
+
+    Get the core node ID for each node of a grid. If a node is not a core
+    node, then use -1.
+
+    Parameters
+    ----------
+    grid : ModelGrid
+        A ModelGrid.
+
+    Returns
+    -------
+    ndarray of int
+        Ids of each of the grid's core nodes.
 
     Examples
     --------
     >>> from landlab import RasterModelGrid
+    >>> from landlab.utils import get_core_node_at_node
+
     >>> grid = RasterModelGrid((4, 5))
+    >>> get_core_node_at_node(grid)
+    array([-1, -1, -1, -1, -1,
+           -1,  0,  1,  2, -1,
+           -1,  3,  4,  5, -1,
+           -1, -1, -1, -1, -1])
+
     >>> grid.status_at_node[13] = grid.BC_NODE_IS_FIXED_VALUE
     >>> grid.status_at_node[2] = grid.BC_NODE_IS_CLOSED
-    >>> matrix_row_at_node(grid)
-    array([-1, -1, -1, -1, -1, -1,  0,  1,  2, -1, -1,  3,  4, -1, -1, -1, -1, -1, -1, -1])
+    >>> get_core_node_at_node(grid)
+    array([-1, -1, -1, -1, -1,
+           -1,  0,  1,  2, -1,
+           -1,  3,  4, -1, -1,
+           -1, -1, -1, -1, -1])
     """
-    matrix_row_at_node = -np.ones(grid.number_of_nodes, dtype=np.int)
-    matrix_row_at_node[grid.core_nodes] = np.arange(
+    core_node_at_node = -np.ones(grid.number_of_nodes, dtype=np.int)
+    core_node_at_node[grid.core_nodes] = np.arange(
         grid.number_of_core_nodes, dtype=np.int
     )
-    return matrix_row_at_node
+    return core_node_at_node
 
 
-def make_core_node_matrix(grid, value, sparse=False):
+def get_matrix_entries(grid, coef_at_link=None):
+    """Get entries of a sparse matrix.
+
+    Parameters
+    ----------
+    grid : RasterModelGrid, HexModelGrid
+        A landlab grid.
+    coef_at_link : ndarray
+        Coefficients at links used to construct the matrix.
+
+    Returns
+    -------
+    tuple of (data, (row_ind, col_inds))
+        Values of matrix elements along with their corresponding row and column
+        index.
     """
-    Construct and return a matrix for the core nodes, plus a right-hand side vector
-    containing values based on the input array `value`.
-
-    Examples
-    --------
-    >>> from landlab import RasterModelGrid
-    >>> grid = RasterModelGrid((4, 5))
-    >>> grid.status_at_node[13] = grid.BC_NODE_IS_FIXED_VALUE
-    >>> grid.status_at_node[2] = grid.BC_NODE_IS_CLOSED
[…]
