### 2020-08-17 1d1b149
add cythonized functions for get_core_node_matrix
