### 2020-08-17 60cfd50
remove/rename cfuncs.pyx
