### 2020-08-18 5eb63f5
add unit tests for get_core_node_matrix
