+++ tests/utils/test_matrix.py
@@ -1,0 +1,225 @@
+import numpy as np
+import pytest
+from numpy.testing import assert_array_equal
+
+from landlab import HexModelGrid, RasterModelGrid
+from landlab.utils import get_core_node_matrix
+
+
+@pytest.mark.parametrize("diff", (-1, 1))
+def test_bad_value_at_node(diff):
+    grid = RasterModelGrid((5, 6))
+    value_at_node = np.full(grid.number_of_nodes + diff, 1.0)
+
+    with pytest.raises(ValueError):
+        get_core_node_matrix(grid, value_at_node)
+
+
+@pytest.mark.parametrize("diff", (-1, 1))
+def test_bad_coef_at_link(diff):
+    grid = RasterModelGrid((5, 6))
+    value_at_node = np.full(grid.number_of_nodes, 1.0)
+    coef_at_link = np.full(grid.number_of_links + diff, 1.0)
+
+    with pytest.raises(ValueError):
+        get_core_node_matrix(grid, value_at_node, coef_at_link=coef_at_link)
+
+
+def test_default_bc():
+    grid = RasterModelGrid((5, 6))
+    value_at_node = np.full(grid.number_of_nodes, 1.0, dtype=float)
+
+    mat, rhs = get_core_node_matrix(grid, value_at_node)
+
+    assert_array_equal(
+        mat.toarray(),
+        [
+            [-4.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
+            [1.0, -4.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
+            [0.0, 1.0, -4.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0],
+            [0.0, 0.0, 1.0, -4.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0],
+            [1.0, 0.0, 0.0, 0.0, -4.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0],
+            [0.0, 1.0, 0.0, 0.0, 1.0, -4.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0],
+            [0.0, 0.0, 1.0, 0.0, 0.0, 1.0, -4.0, 1.0, 0.0, 0.0, 1.0, 0.0],
+            [0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, -4.0, 0.0, 0.0, 0.0, 1.0],
+            [0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, -4.0, 1.0, 0.0, 0.0],
+            [0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, -4.0, 1.0, 0.0],
+            [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, -4.0, 1.0],
+            [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, -4.0],
+        ],
+    )
+    assert_array_equal(
+        rhs,
+        [
+            [-2.0],
+            [-1.0],
+            [-1.0],
+            [-2.0],
+            [-1.0],
+            [0.0],
+            [0.0],
+            [-1.0],
+            [-2.0],
+            [-1.0],
+            [-1.0],
+            [-2.0],
+        ],
+    )
+
+
+def test_small_grids():
[…]
