### 2020-08-18 #2341 @kbarnhart
@pfeiffea I just checked the PDF and found a few formatting things to fix (e.g., reference capitalization, @mcflugen's  affiliation was incorrect). Am fixing it and will update here when done.
