### 2020-08-18 5851110
refactor link_with_angle to use angle_of_link
