+++ joss/in_preparation/network_sediment_transporter/papers.bib
@@ -102,7 +102,7 @@ year = {2019}
 
 @article{Cuietal2006a,
 author = { Yantao   Cui and  Gary   Parker  and  Christian   Braudrick  and  William E   Dietrich  and  Brian   Cluer },
-title = {Dam Removal Express Assessment Models (DREAM).},
+title = {{Dam Removal Express Assessment Models (DREAM). Part 1: Model development and validation}},
 journal = {Journal of Hydraulic Research},
 volume = {44},
 number = {3},
@@ -121,7 +121,7 @@ eprint = {
 
 @article{Cuietal2006b,
 author = { Yantao   Cui  and  Christian   Braudrick  and  William E.   Dietrich  and  Brian   Cluer  and  Gary   Parker },
-title = {Dam Removal Express Assessment Models (DREAM)},
+title = {{Dam Removal Express Assessment Models (DREAM). Part 2: Sample runs/sensitivity tests}},
 journal = {Journal of Hydraulic Research},
 volume = {44},
 number = {3},
