+++ joss/in_preparation/network_sediment_transporter/papers.bib
@@ -1,6 +1,6 @@
 @Article{Hobley2017Creative,
     AUTHOR = {Hobley, D. E. J. and Adams, J. M. and Nudurupati, S. S. and Hutton, E. W. H. and Gasparini, N. M. and Istanbulluoglu, E. and Tucker, G. E.},
-    TITLE = {Creative computing with Landlab: an open-source toolkit for building, coupling, and exploring two-dimensional numerical models of Earth-surface dynamics},
+    TITLE = {{Creative computing with Landlab: an open-source toolkit for building, coupling, and exploring two-dimensional numerical models of Earth-surface dynamics}},
     JOURNAL = {Earth Surface Dynamics},
     VOLUME = {5},
     YEAR = {2017},
@@ -12,7 +12,7 @@
 
 @Article{Barnhart2020Short,
     AUTHOR = {Barnhart, K. R. and Hutton, E. W. H. and Tucker, G. E. and Gasparini, N. M. and Istanbulluoglu, E. and Hobley, D. E. J. and Lyons, N. J. and Mouchene, M. and Nudurupati, S. S. and Adams, J. M. and Bandaragoda, C.},
-    TITLE = {Short communication: Landlab v2.0: A software package for Earth surface dynamics},
+    TITLE = {{Short communication: Landlab v2.0: A software package for Earth surface dynamics}},
     JOURNAL = {Earth Surface Dynamics Discussions},
     VOLUME = {2020},
     YEAR = {2020},
@@ -25,7 +25,6 @@
     abstract = {The movement of sediment from source to sink in a watershed is a complex process with multiple interactions and feedbacks across scales. At small scales, the size characteristics of a sediment mixture affect the transport of that sediment through hiding and through the nonlinear effect of sand content on gravel transport. At large scales, the channel network itself, through network geometry and the spatial pattern of transport capacity, adds additional complexity in organizing sediment moving through the system such that aggradational hotspots can emerge. The purpose of this paper is to present a Lagrangian framework for exploring complexities of mixed-size sediment transport in gravel-bedded river networks. The present model builds off of previous network-based, bed-material sediment transport models; but key advancements presented herein (i) allow for a mixture of sediment sizes in the river network, (ii) incorporate a mixed-size sediment transport equation, and (iii) utilize a daily flow hydrograph to drive intermittent transport. The model is applied to the roughly 4700 km2Methow River Basin in Washington State, USA, using simplified model inputs with the goal of illustrating the utility of the model for motivating future work.},
     author = {Czuba, Jonathan A.},
     doi = {10.1016/j.geomorph.2018.08.031},
-    file = {:C$\backslash$:/Users/pfeif/Documents/Papers/Czuba-2018-A lagrangian framework for exploring the complexities of mixed-size sediment transport- Geomorphology.pdf:pdf},
     issn = {0169555X},
     journal = {Geomorphology},
     keywords = {Complexity,Gravel-bedded river,Network model,River network,Sediment mixture,Sediment transport},
@@ -40,7 +39,6 @@
   @article{An2017,
     author = {An, Chenge and Cui, Yantao and Fu, Xudong and Parker, Gary},
     doi = {10.1002/esp.4195},
-    file = {:C$\backslash$:/Users/pfeif/Documents/Papers/An-2017-Gravel bed river evolution in earthquake prone regions subject to cycled hydrographs and repeated sediment pulses-ESPL.pdf:pdf},
     issn = {10969837},
     journal = {Earth Surface Processes and Landforms},
[…]
