### 2020-08-18 c5c79b2
use enum for link directions
