+++ landlab/grid/hex_mappers.py
@@ -9,22 +9,25 @@ Mapping functions unique to hex grids
 
     ~landlab.grid.hex_mappers.map_link_vector_components_to_node_hex
 """
+import enum
 
 import numpy as np
 
-LINK_TO_EAST = 0  # array column of link to east/right of node, horizontal grid
-LINK_TO_NNE = 1  # array column of link to nne/upper right of node, horizontal grid
-LINK_TO_NNW = 2  # array column of link to nnw/upper left of node, horizontal grid
-LINK_TO_WEST = 3  # array column of link to west/left of node, horizontal grid
-LINK_TO_SSW = 4  # array column of link to ssw/lower left of node, horizontal grid
-LINK_TO_SSE = 5  # array column of link to sse/lower right of node, horizontal grid
 
-LINK_TO_ENE = 0  # array column of link to ene/upper right of node, vertical grid
-LINK_TO_NORTH = 1  # array column of link to north/top of node, vertical grid
-LINK_TO_WNW = 2  # array column of link to wnw/upper left of node, vertical grid
-LINK_TO_WSW = 3  # array column of link to wsw/lower left of node, vertical grid
-LINK_TO_SOUTH = 4  # array column of link to south/bottom of node, vertical grid
-LINK_TO_ESE = 5  # array column of link to ese/lower right of node, vertical grid
+class LinkAtNode(enum.IntEnum):
+    EAST = 0  # array column of link to east/right of node, horizontal grid
+    ENE = 0  # array column of link to ene/upper right of node, vertical grid
+    NNE = 1  # array column of link to nne/upper right of node, horizontal grid
+    NORTH = 1  # array column of link to north/top of node, vertical grid
+    NNW = 2  # array column of link to nnw/upper left of node, horizontal grid
+    WNW = 2  # array column of link to wnw/upper left of node, vertical grid
+    WEST = 3  # array column of link to west/left of node, horizontal grid
+    WSW = 3  # array column of link to wsw/lower left of node, vertical grid
+    SSW = 4  # array column of link to ssw/lower left of node, horizontal grid
+    SOUTH = 4  # array column of link to south/bottom of node, vertical grid
+    SSE = 5  # array column of link to sse/lower right of node, horizontal grid
+    ESE = 5  # array column of link to ese/lower right of node, vertical grid
+
 
 SIN60 = np.sin(np.deg2rad(60.0))
 
@@ -41,11 +44,13 @@ def map_link_vector_components_to_node_hex(grid, data_at_link):
 
     Examples
     --------
+    >>> import numpy as np
     >>> from landlab import HexModelGrid
     >>> from landlab.grid.mappers import map_link_vector_components_to_node
     >>> grid = HexModelGrid((3, 3))
     >>> link_data = np.zeros(grid.number_of_links) + 0.5 * 3.0**0.5
-    >>> link_data[grid.link_with_angle(0.0)] = 0.0
+    >>> link_data[np.isclose(grid.angle_of_link, 0.0)] = 0.0
+    >>> # link_data[grid.link_with_angle(0.0)] = 0.0
     >>> vx, vy = map_link_vector_components_to_node(grid, link_data)
     >>> vx
     array([ 0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.])
[…]
