### 2020-08-18 #2341 @kbarnhart
OK @pfeiffea I've fixed the small formatting things (was at landlab/landlab/pull/1235)
