### 2020-08-20 c675be8
use grid.angle_of_link
