### 2020-08-27 #2341 @kthyng
@pfeiffea Looks like you've addressed reviewer comments. I'd like to check in with them.

@zsylvester, @ebgoldstein: Please see comments from @pfeiffea above and let us know if you have more comments or if you are satisfied in the responses. Thanks!
