### 2020-08-27 #2341 @ebgoldstein
the responses, paper changes, and tutorial notebook changes all look good to me. I have no more comments.
