### 2020-08-28 #2341 @zsylvester
It all looks great! Now I understand better the differences between this model and the one in Czuba (2018). Only a few minor comments:

- The 'Repository' link on the first page links to the Landlab repo; wouldn't it be better if it linked to [NST](https://github.com/landlab/landlab/tree/master/landlab/components/network_sediment_transporter)?
- There is no URL associated with 'Archive' (page 1)
- 'open-source Python package' instead of 'Open Source Python package' (page 2)
- capitalize 'python' in 'Landlab python package' (page 3)
