### 2020-08-31 519a2df
test one approach to ED
