### 2020-08-31 #2341 @kbarnhart
I've addressed the second two comments identified by @zsylvester on landlab/landlab/pull/1237. Should get merged later today after CI runs.

My answers to the first two are as one of the authors of this paper but reflect knowledge that I have of the JOSS pdf-building and review process that come from being a topical editor.

* I'm not sure whether its possible to point the repository link to
  https://github.com/landlab/landlab/tree/master/landlab/components/network_sediment_transporter
  instead of to
  https://github.com/landlab/landlab
  (probably a question for @arfon?)

* I don't think the Archive link should work yet because we haven't yet made an archive (waiting until the end) and no archive has been set.
