### 2020-09-01 #2341 @kbarnhart
Updates addressing half of @zsylvester 's comments now merged into master.
