### 2020-09-08 #2341 @pfeiffea
@kthyng let us know if these changes adequately address @zsylvester 's comments.

If so, what are our next steps?
