+++ notebooks/tutorials/geohealth/data/sample_locations.txt
@@ -1,0 +1,27 @@
+FID,Count,System,Sample,SampleID,Latitude,Longitude,GeoNetwork
+0,2.000000000000000,D_1,Res,D_1_Res,18.033149999999999,-65.981430000000003,Moved 200m downstream to link upstream of D_2
+1,14.000000000000000,D_2,Res,D_2_Res,18.030944999999999,-65.980138999999994,No change
+2,24.000000000000000,F,Res,F_Res,18.033158000000000,-65.988489999999999,Moved downhill to north network 300 meters
+3,37.000000000000000,B,URes,B_URes,18.077463000000002,-66.051862999999997,Moved douth 100 m - uncertain on side of ridge
+4,38.000000000000000,B,LRes,B_LRes,18.076606000000002,-66.053433999999996,Moved 100m north onto network
+5,50.000000000000000,A,Res,A_Res,18.063904000000001,-66.029454999999999,No change
+6,63.000000000000000,AB,1,AB_1,18.259383000000000,-66.103465999999997,No change
+7,64.000000000000000,AB,2,AB_2,18.258465000000001,-66.094181000000006,No change
+8,65.000000000000000,ABCG,3,ABCG_3,18.256978000000000,-66.065113999999994,No change
+9,66.000000000000000,CG,4,CG_4,18.259920999999999,-66.026111999999998,No change
+10,67.000000000000000,CG,5,CG_5,18.259032999999999,-66.022549999999995,No change
+11,68.000000000000000,AW,1,AW_1,18.013596000000000,-65.986390000000000,Moved north 100m to network
+12,69.000000000000000,AW,2,AW_2,18.001303000000000,-65.993511999999996,Moved north 100m to network
+13,70.000000000000000,AW,3,AW_3,17.993803000000000,-66.008157999999995,Moved east 100m to network
+14,71.000000000000000,AW,4,AW_4,17.992594000000000,-66.008003000000002,Moved south 100m to network
+15,79.000000000000000,PLW,E1,PLW_E1,18.050659000000000,-66.010600999999994,No change
+16,80.000000000000000,PLW,E2,PLW_E2,18.037094000000000,-66.006399999999999,No change
+17,81.000000000000000,PLW,E3,PLW_E3,18.028898000000002,-66.011561999999998,Moved 100m eat to network
+18,82.000000000000000,PLW,W1,PLW_W1,18.079107000000000,-66.053837000000001,No change
+19,83.000000000000000,PLW,W2,PLW_W2,18.072540000000000,-66.056428999999994,Moved 100m west to network
+20,84.000000000000000,PLW,W3,PLW_W3,18.063388000000000,-66.051159999999996,Moved 100m north to nework
+21,85.000000000000000,PLW,W4,PLW_W4,18.047870000000000,-66.037969000000004,Moved 50 m north to network
+22,86.000000000000000,PLW,W5,PLW_W5,18.041748999999999,-66.034698000000006,Moved 100 m west to network
+23,87.000000000000000,PLW,W6,PLW_W6,18.031392000000000,-66.031881999999996,No change
+24,88.000000000000000,PLW,W7,PLW_W7,18.025971999999999,-66.031214000000006,Moved 100m northwest to west bank drainage
+25,89.000000000000000,PLW,L,PLW_L,18.018623000000002,-66.017386000000002,Moved 100 m southeast to south back of lago
