+++ landlab/components/lake_fill/debug_lake_fill_tests.py
@@ -1,0 +1,375 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Sun Sep 13 10:08:51 2020
+
+@author: gtucker
+"""
+
+import numpy as np
+from landlab import RasterModelGrid
+from landlab.components import LakeMapperBarnes, FlowAccumulator
+from landlab.components import FlowDirectorSteepest
+from collections import deque
+
+
+mg = RasterModelGrid((5, 6), xy_spacing=2.)
+for edge in ('left', 'top', 'bottom'):
+    mg.status_at_node[mg.nodes_at_edge(edge)] = mg.BC_NODE_IS_CLOSED
+z = mg.add_zeros("topographic__elevation", at="node", dtype=float)
+z.reshape(mg.shape)[2, 1:-1] = [2., 1., 0.5, 1.5]
+z.reshape(mg.shape)[1, 1:-1] = [2.1, 1.1, 0.6, 1.6]
+z.reshape(mg.shape)[3, 1:-1] = [2.2, 1.2, 0.7, 1.7]
+z_init = z.copy()
+fa = FlowAccumulator(mg)
+lmb = LakeMapperBarnes(mg, method='Steepest', surface=z_init, 
+                       fill_flat=False,
+                       redirect_flow_steepest_descent=False,
+                       track_lakes=False)
+
+lmb.run_one_step()
+z_out = np.array([ 0. ,  0. ,  0. ,  0. ,  0. ,  0. ,
+                   0. ,  2.1,  1.5,  1.5,  1.6,  0. ,
+                   0. ,  2. ,  1.5,  1.5,  1.5,  0. ,
+                   0. ,  2.2,  1.5,  1.5,  1.7,  0. ,
+                   0. ,  0. ,  0. ,  0. ,  0. ,  0. ])
+assert(np.allclose(z, z_out))
+#        True
+assert(not np.all(np.equal(z, z_out)))  # those 1.5's are actually a bit > 1.5
+#         False
+try:
+    lmb.lake_map  # not created, as we aren't tracking
+except ValueError:
+    print('ValueError was raised: ' +
+          'Enable tracking to access information about lakes')
+#         ValueError was raised: Enable tracking to access information about lakes
+assert(not lmb.was_there_overfill)  # everything fine with slope adding
+#         False
+
+fd = FlowDirectorSteepest(mg)
+fa = FlowAccumulator(mg)  # routing will work fine now
+fd.run_one_step()
+fa.run_one_step()
+assert(np.all(mg.at_node['flow__sink_flag'][mg.core_nodes] == 0))
+#         True
+drainage_area = np.array([  0.,   0.,   0.,   0.,   0.,   0.,
+              0.,   4.,   8.,  12.,   4.,   4.,
+              0.,   4.,  16.,  36.,  40.,  40.,
+              0.,   4.,   8.,   4.,   4.,   4.,
+              0.,   0.,   0.,   0.,   0.,   0.])
+assert(np.allclose(mg.at_node['drainage_area'], drainage_area))
+#         True
+
+#         Test two pits:
+
+z[:] = mg.node_x.max() - mg.node_x
+z[23] = 1.3
+z[15] = 0.3
+z[10] = 1.3  # raise "guard" exit nodes
+z[7] = 2.  # is a lake on its own, if D8
+z[9] = 0.5
+z[14] = 0.6  # [9, 14, 15] is a lake in both methods
+z[16] = 1.2
+z[22] = 0.9  # a non-contiguous lake node also draining to 16 if D8
+z_init = z.copy()
+fa = FlowAccumulator(mg)
+lmb = LakeMapperBarnes(mg, method='D8', fill_flat=True,
+        track_lakes=True)
+lmb.run_one_step()  # note the D8 routing now
[…]
