+++ landlab/components/lake_fill/lake_fill_barnes.py
@@ -1216,19 +1216,19 @@ class LakeMapperBarnes(Component):
             # print(type(np.array(self._steepestslopes)))
             # print(np.array(self._steepestslopes).shape)
             has_diags = len(self._neighbor_arrays) - 1  # 0 or 1
-            if has_diags:
-                print('Num links: ' + str(self._grid.number_of_links))
-                print('Num diags: ' + str(self._neighbor_arrays[1].shape))
-                print('Diag array:')
-                print(self._neighbor_arrays[1])
-                print('liminal nodes:')
-                print(liminal_nodes)
+            # if has_diags:
+            #     print('Num links: ' + str(self._grid.number_of_links))
+            #     print('Num diags: ' + str(self._neighbor_arrays[1].shape))
+            #     print('Diag array:')
+            #     print(self._neighbor_arrays[1])
+            #     print('liminal nodes:')
+            #     print(liminal_nodes)
             from .cfuncs import redirect_dirs
             redirect_dirs(
                             np.array(liminal_nodes),
                             self._neighbor_arrays[0],
-                            self._neighbor_arrays[has_diags],
                             self._link_arrays[0],
+                            self._neighbor_arrays[has_diags],
                             self._link_arrays[has_diags],
                             self._grid.status_at_node,
                             int(self._grid.BC_NODE_IS_CLOSED),
@@ -1239,15 +1239,19 @@ class LakeMapperBarnes(Component):
                             self._steepestslopes,
                             has_diags,
                             )
-            for liminal in []: #liminal_nodes:
+            for liminal in []: # liminal_nodes:
                 min_elev = LARGE_ELEV
                 min_link = -1
+                # debug = 0
                 for neighbor_set, link_set in zip(
                     self._neighbor_arrays, self._link_arrays
                 ):
                     #print(type(neighbor_set), liminal)
                     #print(neighbor_set.shape, link_set.shape)
-                    #print(neighbor_set)
+                    # if has_diags:
+                    #     print('\n pass ' + str(debug))
+                    #     debug += 1
+                    #     print(neighbor_set)
                     neighbors = neighbor_set[liminal]
                     #print(neighbors)
                     neighbors_valid = np.not_equal(neighbors, -1)
@@ -1683,7 +1687,7 @@ class LakeMapperBarnes(Component):
         ensures the information about the lake and the water surface
         topography are all updated cleanly and correctly.
         """
-        print('lmb ros')
+        #print('lmb ros')
         import time
         s = time.time()
         if "flow__receiver_node" in self._grid.at_node:
@@ -1717,7 +1721,7 @@ class LakeMapperBarnes(Component):
         #print(e-s)
         s = e
         if self._track_lakes:
-            print('track_lakes is True')
+            #print('track_lakes is True')
             for edgenode in self._edges:
                 _open.add_task(edgenode, priority=self._surface[edgenode])
             e = time.time()
@@ -1750,7 +1754,7 @@ class LakeMapperBarnes(Component):
                     _, _ = self._fa.accumulate_flow(update_flow_director=False)
 
         else:  # not tracked
-            print('Track lakes False')
+            #print('Track lakes False')
             # note we've already checked _dontredirect is True in setup,
             # so we don't need to worry about these cases.
             for edgenode in self._edges:
