+++ landlab/io/obj.py
@@ -15,18 +15,32 @@ import os
 def _write_nodes_as_vertices(file, grid, z):
     """Write node (x,y,z) coordinates as OBJ vertices."""
     for i in range(grid.number_of_nodes):
-        file.write('v ' + str(grid.x_of_node[i]) + ' '
-                   + str(grid.y_of_node[i]) + ' '
-                   + str(z[i]) + '\n')
+        file.write(
+            "v "
+            + str(grid.x_of_node[i])
+            + " "
+            + str(grid.y_of_node[i])
+            + " "
+            + str(z[i])
+            + "\n"
+        )
+
 
 def _write_triangular_patches_as_faces(file, grid, z):
     """Write triangular patches as OBJ faces.
 
     OBJ format uses vertex/UVtexture/normal; latter two left blank here."""
     for i in range(len(grid.nodes_at_patch)):
-        file.write('f ' + str(grid.nodes_at_patch[i,0] + 1) + '// '
-                   + str(grid.nodes_at_patch[i,1] + 1) + '// '
-                   + str(grid.nodes_at_patch[i,2] + 1) + '//\n')
+        file.write(
+            "f "
+            + str(grid.nodes_at_patch[i, 0] + 1)
+            + "// "
+            + str(grid.nodes_at_patch[i, 1] + 1)
+            + "// "
+            + str(grid.nodes_at_patch[i, 2] + 1)
+            + "//\n"
+        )
+
 
 def _write_quad_patches_as_triangular_faces(file, grid, z):
     """Write quad patches as OBJ (triangle) faces.
@@ -36,22 +50,37 @@ def _write_quad_patches_as_triangular_faces(file, grid, z):
 
     OBJ format uses vertex/UVtexture/normal; latter two left blank here."""
     for i in range(len(grid.nodes_at_patch)):
-        file.write('f ' + str(grid.nodes_at_patch[i,0] + 1) + '// '
-                   + str(grid.nodes_at_patch[i,1] + 1) + '// '
-                   + str(grid.nodes_at_patch[i,2] + 1) + '//\n')
-        file.write('f ' + str(grid.nodes_at_patch[i,2] + 1) + '// '
-                   + str(grid.nodes_at_patch[i,3] + 1) + '// '
-                   + str(grid.nodes_at_patch[i,0] + 1) + '//\n')
+        file.write(
+            "f "
+            + str(grid.nodes_at_patch[i, 0] + 1)
+            + "// "
+            + str(grid.nodes_at_patch[i, 1] + 1)
+            + "// "
+            + str(grid.nodes_at_patch[i, 2] + 1)
+            + "//\n"
+        )
+        file.write(
+            "f "
+            + str(grid.nodes_at_patch[i, 2] + 1)
+            + "// "
+            + str(grid.nodes_at_patch[i, 3] + 1)
+            + "// "
+            + str(grid.nodes_at_patch[i, 0] + 1)
+            + "//\n"
+        )
+
 
 def _patches_are_triangles(grid):
     """Returns True if patches have 3 nodes, False otherwise."""
     return grid.nodes_at_patch.shape[1] == 3
 
+
 def _patches_are_quads(grid):
     """Returns True if patches have 4 nodes, False otherwise."""
     return grid.nodes_at_patch.shape[1] == 4
 
-def write_obj(path, grid, field_for_z='topographic__elevation', clobber=False):
+
+def write_obj(path, grid, field_for_z="topographic__elevation", clobber=False):
     """Write landlab grid and one field to Wavefront OBJ.
 
     Parameters
@@ -94,10 +123,10 @@ def write_obj(path, grid, field_for_z='topographic__elevation', clobber=False):
         raise ValueError("file exists")
 
[…]
