### 2020-09-21 #2341 @kthyng
Paper looks great!

Next, @pfeiffea can you verify the version of software you want tagged with this publication?

After that, please upload an archive of your software to something like Zenodo. When doing so, please make sure that the title of author list of the archive exactly match your JOSS paper. Then, report back here with the doi.
