### 2020-09-22 47fb34f
add write_obj to __init__
