+++ landlab/io/__init__.py
@@ -13,12 +13,14 @@ from .esri_ascii import (
     write_esri_ascii,
 )
 from .shapefile import read_shapefile
+from .obj import write_obj
 
 __all__ = [
     "read_esri_ascii",
     "read_asc_header",
     "read_shapefile",
     "write_esri_ascii",
+    "write_obj",
     "MissingRequiredKeyError",
     "KeyTypeError",
     "DataSizeError",
