### 2020-09-23 #2341 @pfeiffea
Thanks, @kthyng!

Landlab v2.2.0

Here's the doi for a Zenodo archive: 10.5281/zenodo.4044683
