### 2020-09-23 94f34ee
add docstrings; write to path or file-like
