+++ landlab/io/obj.py
@@ -8,66 +8,58 @@ OBJ functions
 
     ~landlab.io.obj.write_obj
 """
-
 import os
+import pathlib
 
 
-def _write_nodes_as_vertices(file, grid, z):
-    """Write node (x,y,z) coordinates as OBJ vertices."""
-    for i in range(grid.number_of_nodes):
-        file.write(
-            "v "
-            + str(grid.x_of_node[i])
-            + " "
-            + str(grid.y_of_node[i])
-            + " "
-            + str(z[i])
-            + "\n"
-        )
+def _write_obj_vertices(file_like, xyz_at_vertex):
+    """Write node (x,y,z) coordinates as OBJ vertices.
+
+    Parameters
+    ----------
+    file_like : file-like
+        Opened file-like object to write vertices to.
+    xyz_at_vertex : array-like, shape *(n_vertices, 3)*
+        (x, y, z) values for each vertex.
+    """
+    for x, y, z in xyz_at_vertex:
+        print(f"v {x} {y} {z}", file=file_like)
 
 
-def _write_triangular_patches_as_faces(file, grid, z):
+def _write_triangles_as_obj_faces(file_like, vertices_at_face):
     """Write triangular patches as OBJ faces.
 
-    OBJ format uses vertex/UVtexture/normal; latter two left blank here."""
-    for i in range(len(grid.nodes_at_patch)):
-        file.write(
-            "f "
-            + str(grid.nodes_at_patch[i, 0] + 1)
-            + "// "
-            + str(grid.nodes_at_patch[i, 1] + 1)
-            + "// "
-            + str(grid.nodes_at_patch[i, 2] + 1)
-            + "//\n"
-        )
+    OBJ format uses vertex/UVtexture/normal; latter two left blank here.
+
+    Parameters
+    ----------
+    file_like : file-like
+        Opened file-like object to write faces to.
+    vertices_at_face : array-like of int, shape *(n_faces, 3)*
+        Vertex ids for each triangle.
+    """
+    for vertex_0, vertex_1, vertex_2 in vertices_at_face:
+        print(f"f {vertex_0}// {vertex_1}// {vertex_2}//", file=file_like)
 
 
-def _write_quad_patches_as_triangular_faces(file, grid, z):
+def _write_quads_as_obj_faces(file_, vertices_at_face):
     """Write quad patches as OBJ (triangle) faces.
 
     Subdivide each patch into two triangles using vertices 0-1-2 (upper left)
     and 2-3-0 (lower right).
 
-    OBJ format uses vertex/UVtexture/normal; latter two left blank here."""
-    for i in range(len(grid.nodes_at_patch)):
-        file.write(
-            "f "
-            + str(grid.nodes_at_patch[i, 0] + 1)
-            + "// "
-            + str(grid.nodes_at_patch[i, 1] + 1)
-            + "// "
-            + str(grid.nodes_at_patch[i, 2] + 1)
-            + "//\n"
-        )
-        file.write(
-            "f "
-            + str(grid.nodes_at_patch[i, 2] + 1)
-            + "// "
-            + str(grid.nodes_at_patch[i, 3] + 1)
-            + "// "
-            + str(grid.nodes_at_patch[i, 0] + 1)
-            + "//\n"
-        )
+    OBJ format uses vertex/UVtexture/normal; latter two left blank here.
+
+    Parameters
+    ----------
+    file_like : file-like
+        Opened file-like object to write faces to.
+    vertices_at_face : array-like of int, shape *(n_faces, 4)*
+        Vertex ids for each quadrilateral.
+    """
[…]
