### 2020-09-23 0900f94
add more unit tests for obj writer
